% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTING.R
\docType{class}
\name{Account-class}
\alias{Account-class}
\alias{Account}
\alias{Account<-}
\alias{Account<-,numeric-method}
\alias{Account,array-method}
\title{An S4 class to represent a bank account}
\usage{
Account(x)

Account(x) <- value

\S4method{Account}{numeric}(x) <- value

\S4method{Account}{array}(x)
}
\description{
An S4 class to represent a bank account
}
\details{
Additional details...
}
\section{Methods (by generic)}{
\itemize{
\item \code{Account(numeric) <- value}: description

\item \code{Account(array)}: description

}}
\section{Functions}{
\itemize{
\item \code{Account()}: description

\item \code{Account(x) <- value}: description

}}
\section{Slots}{

\describe{
\item{\code{balance}}{A length-one numeric vector}
}}

