% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Boxplot.R
\docType{class}
\name{Boxplot-class}
\alias{Boxplot-class}
\alias{show,Boxplot-method}
\title{S4 class Boxplot}
\usage{
\S4method{show}{Boxplot}(object)
}
\arguments{
\item{object}{An object of class \code{\link[=Boxplot-class]{Boxplot-class()}}}
}
\description{
Slots can be accessed and assigned using functions corresponding to slot name.
See \verb{See Also} section below.
}
\details{
Objects of class \code{Boxplot} are created with \code{Boxplot()}
\subsection{Multi-Language Support}{

Text with multi-language supported can be provided as a named list. Available languages:
\itemize{
\item \code{en}: English (default)
\item \code{es}: Spanish
\item \code{fr}: French
}
}

\subsection{Note}{

Character strings in \code{Code}, \code{Label}, and \code{Description} must all be same length
as the number of performance indicators (\code{nPIs}) in \code{Value}
}
}
\section{Slots}{

\describe{
\item{\code{Code}}{A \emph{short} code for the Performance Indicators for this object.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{\code{Label}}{A short label for the Performance Indicators for this object. Used to label axes on charts.
Can be longer than \code{Code} but recommended to keep short as possible so it shows clearly in plots and tables.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{\code{Description}}{A description for the Performance Indicators for this object.
Can include Markdown, see \code{Examples}.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{\code{Value}}{A numeric array with the stochastic performance indicator values for each
simulation (sim), operating model (OM), management procedure (MP), and performance indicator (PI).
Dimensions: c(\code{nsim}, \code{nOM}, \code{nMP}, and \code{nPI})}

\item{\code{Preset}}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. See \code{Examples}}
}}

\examples{
boxplot <- Boxplot(Code=c('PI1', 'PI2', 'PI3'),
                   Label=c('Performance Indicator 1',
                         'Performance Indicator 2',
                         'Performance Indicator 3'),
                   Description = c('This is the description for PI 1',
                                   'This is the description for PI 2',
                                   'This is the description for PI 3'),
                   Value=array(runif(3000), dim=c(10, 10, 10,3)),
                   Preset=list('PI and P2'=1:2)

)

boxplot

Code(boxplot)
Metadata(boxplot)

# Value(boxplot)
Value(boxplot) <- array()

Preset(boxplot)
Preset(boxplot) <- list()

boxplot

}
\seealso{
\code{\link[=Boxplot-methods]{Boxplot-methods()}}, \code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}},
\code{\link[=Metadata]{Metadata()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}}
}
