% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Boxplot.R, R/class_Slick.R
\docType{methods}
\name{Boxplot}
\alias{Boxplot}
\alias{Boxplot<-}
\title{Methods for Creating, Accessing and Assigning \code{Boxplot} objects}
\usage{
Boxplot(Code, Label, Description, Value, Preset)

Boxplot(Slick) <- value
}
\arguments{
\item{Code}{A \emph{short} code for the Performance Indicators for this object.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Label}{A short label for the Performance Indicators for this object. Used to label axes on charts.
Can be longer than \code{Code} but recommended to keep short as possible so it shows clearly in plots and tables.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Description}{A description for the Performance Indicators for this object.
Can include Markdown, see \code{Examples}.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Value}{A numeric array with the stochastic performance indicator values for each
simulation (sim), operating model (OM), management procedure (MP), and performance indicator (PI).
Dimensions: c(\code{nsim}, \code{nOM}, \code{nMP}, and \code{nPI})}

\item{Preset}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. See \code{Examples}}

\item{Slick}{An object of class \code{\link[=Slick-class]{Slick-class()}}}

\item{value}{An object of class \code{\link[=Boxplot-class]{Boxplot-class()}}}
}
\description{
An object of class \code{Boxplot} contains information for the Boxplot chart.
The \code{Boxplot} function is used both to create and modify an \code{\link[=Boxplot-class]{Boxplot-class()}} object.
and to access and assign \code{Boxplot} for an object of class \code{\link[=Slick-class]{Slick-class()}}.
See \code{Details}.
}
\details{
\itemize{
\item \code{Boxplot()}: creates an empty \code{Boxplot} object
\item \code{Boxplot(Code, ...)}: creates a populated \code{Boxplot} object
\item \code{Boxplot(Slick)}: returns the \code{Boxplot} from a \code{Slick} object
\item \code{Boxplot(Slick) <- Boxplot}: assigns a \code{Boxplot} object to a \code{Slick} object
}

Use the  \code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}} functions to access and assign the values for an
existing \code{Boxplot} object, see \code{Examples}
}
\examples{
boxplot <- Boxplot(Code=c('PI1', 'PI2', 'PI3'),
                   Label=c('Performance Indicator 1',
                         'Performance Indicator 2',
                         'Performance Indicator 3'),
                   Description = c('This is the description for PI 1',
                                   'This is the description for PI 2',
                                   'This is the description for PI 3'),
                   Value=array(runif(3000), dim=c(10, 10, 10,3)),
                   Preset=list('PI and P2'=1:2)

)

boxplot

Code(boxplot)
Metadata(boxplot)

# Value(boxplot)
Value(boxplot) <- array()

Preset(boxplot)
Preset(boxplot) <- list()

boxplot

}
\seealso{
\code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Metadata]{Metadata()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}}
}
