% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Code.R, R/fct_Description.R, R/fct_Label.R
\name{Code}
\alias{Code}
\alias{Code<-}
\alias{Description}
\alias{Description<-}
\alias{Label}
\alias{Label<-}
\title{Access or assign \code{Code}, \code{Label}, and \code{Description} for a valid object class}
\usage{
Code(object, lang = "en", type = NULL)

Code(object, type = NULL) <- value

Description(object, lang = "en", type = NULL)

Description(object, type = NULL) <- value

Label(object, lang = "en", type = NULL)

Label(object, type = NULL) <- value
}
\arguments{
\item{object}{An object of class \code{\link[=MPs-class]{MPs-class()}}, \code{\link[=Boxplot-class]{Boxplot-class()}},
\code{\link[=Kobe-class]{Kobe-class()}}, \code{\link[=Quilt-class]{Quilt-class()}}, \code{\link[=Spider-class]{Spider-class()}},
\code{\link[=Timeseries-class]{Timeseries-class()}}, \code{\link[=Tradeoff-class]{Tradeoff-class()}}, or \code{\link[=Slick-class]{Slick-class()}}}

\item{lang}{Optional text string specifying the language (if available).
Either 'en', 'es', or 'fr' for English, Spanish, or French respectively}

\item{type}{Required if \code{object} is class \code{\link[=Slick-class]{Slick-class()}}. One of c('MPs',  'Boxplot',
'Kobe','Quilt', 'Spider', 'Timeseries', 'Tradeoff')}

\item{value}{The replacement value for \code{object}.
See the documentation in the corresponding \code{object} class for details.}
}
\description{
Access or assign \code{Code}, \code{Label}, and \code{Description} for a valid object class
}
\details{
\emph{Note}:
If \code{object} is class \code{\link[=Slick-class]{Slick-class()}}, then \code{type} must be specified.
}
\section{Functions}{
\itemize{
\item \code{Code(object, type = NULL) <- value}: Assign a replacement value for \code{Code} in \code{object}.

\item \code{Description()}: Access \code{Description} from a valid object class

\item \code{Description(object, type = NULL) <- value}: Assign a replacement value for \code{Description} in \code{object}.

\item \code{Label()}: Access \code{Label} from a valid object class

\item \code{Label(object, type = NULL) <- value}: Assign a replacement value for \code{Label} in \code{object}.

}}
\examples{
boxplot <- Boxplot()

Code(boxplot) <- c('PI1', 'PI2', 'PI3')
Label(boxplot) <- c('Performance Indicator 1',
                    'Performance Indicator 2',
                    'Performance Indicator 3')
Description(boxplot) <- c('This is the description for PI 1',
                          'This is the description for PI 2',
                          'This is the description for PI 3')


Code(boxplot)
Label(boxplot)
Description(boxplot)

# Slick Object
mySlick <- Slick()
Code(mySlick, 'Boxplot') <- Code(boxplot)
Label(mySlick, 'Boxplot') <- Label(boxplot)
Description(mySlick, 'Boxplot') <- Description(boxplot)

Code(mySlick, type='Boxplot')
Label(mySlick, type='Boxplot')
Description(mySlick, type='Boxplot')

# Multi-Language
Description(boxplot) <- list(en=c('This is the English description for PI 1',
                                  'This is the English description for PI 2',
                                  'This is the English description for PI 3'),
                             es=c('This is the Spanish description for PI 1',
                                  'This is the Spanish description for PI 2',
                                  'This is the Spanish description for PI 3'),
                             fr=c('This is the French description for PI 1',
                                  'This is the French description for PI 2',
                                  'This is the French description for PI 3')
)

Description(boxplot)
Description(boxplot, 'es')
Description(boxplot, 'fr')
}
\seealso{
\code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=MPs-class]{MPs-class()}}, \code{\link[=Boxplot-class]{Boxplot-class()}},
\code{\link[=Kobe-class]{Kobe-class()}}, \code{\link[=Quilt-class]{Quilt-class()}}, \code{\link[=Spider-class]{Spider-class()}},
\code{\link[=Timeseries-class]{Timeseries-class()}}, \code{\link[=Tradeoff-class]{Tradeoff-class()}}, \code{\link[=Slick-class]{Slick-class()}}
}
