% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_MPs.R, R/class_Quilt.R
\name{Color}
\alias{Color}
\alias{Color<-}
\alias{Color,MPs-method}
\alias{Color<-,MPs-method}
\alias{Color,Quilt-method}
\alias{Color<-,Quilt-method}
\title{Access or assign \code{Color} for \code{MPs} and \code{Quilt} objects}
\usage{
Color(object)

Color(object) <- value

\S4method{Color}{MPs}(object)

\S4method{Color}{MPs}(object) <- value

\S4method{Color}{Quilt}(object)

\S4method{Color}{Quilt}(object) <- value
}
\arguments{
\item{object}{An \code{\link[=MPs-class]{MPs-class()}} or \code{\link[=Quilt-class]{Quilt-class()}} object}

\item{value}{A character vector formatted to match the class of \code{object}. See the documentation for
corresponding \code{object} class for more details.}
}
\value{
Returns a character string with the contents of the \code{Color} slot of
\code{object}
}
\description{
Access or assign \code{Color} for \code{MPs} and \code{Quilt} objects
}
\section{Methods (by class)}{
\itemize{
\item \code{Color(MPs)}: Return \code{Color} from a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Color(MPs) <- value}: Preset Assign \code{Color} to a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Color(Quilt)}: Return \code{Color} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Color(Quilt) <- value}: Preset Assign \code{Color} to a \code{\link[=Quilt-class]{Quilt-class()}} object

}}
