% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_colors.R
\name{Colors}
\alias{Colors}
\alias{Colors<-}
\title{Assign or access the \code{Colors} slot in an object of class \code{Quilt}}
\usage{
Colors(Quilt)

Colors(Quilt) <- value
}
\arguments{
\item{Quilt}{An object of class \code{\link[=Quilt]{Quilt()}}}

\item{value}{A character vector of length 2 specifying the colors for
the maximum and minimum values in the Quilt chart}
}
\description{
Assign or access the \code{Colors} slot in an object of class \code{Quilt}
}
\examples{
quilt <- Quilt()
Colors(quilt) <- c('red', 'blue')
Colors(quilt)

}
