% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{default_mp_colors}
\alias{default_mp_colors}
\title{Set Default Colors for the MPs}
\usage{
default_mp_colors(nMP)
}
\arguments{
\item{nMP}{The number of management procedures}
}
\value{
A character vector of length \code{nMP} with color hex codes
}
\description{
Requires the \code{colorspace} package
}
\examples{
cols <- default_mp_colors(4)
cols

}
