% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Boxplot.R,
%   R/class_Kobe.R
\name{Defaults}
\alias{Defaults}
\alias{Defaults<-}
\alias{Defaults,Boxplot-method}
\alias{Defaults<-,Boxplot-method}
\alias{Defaults,Kobe-method}
\alias{Defaults<-,Kobe-method}
\title{Access or assign the default selections for the plots in the \code{App()}}
\usage{
Defaults(object)

Defaults(object) <- value

\S4method{Defaults}{Boxplot}(object)

\S4method{Defaults}{Boxplot}(object) <- value

\S4method{Defaults}{Kobe}(object)

\S4method{Defaults}{Kobe}(object) <- value
}
\arguments{
\item{object}{A \code{\link[=Boxplot]{Boxplot()}} or \code{\link[=Kobe]{Kobe()}} object}

\item{value}{A list of default selections for the \code{\link[=App]{App()}}}
}
\value{
Returns a list object with default selections from the \code{Defaults} slot of
\code{\link[=Boxplot]{Boxplot()}} and \code{\link[=Kobe]{Kobe()}} objects.
}
\description{
Access or assign the default selections for the plots in the \code{App()}
}
\section{Methods (by class)}{
\itemize{
\item \code{Defaults(Boxplot)}: Defaults \code{Code} from a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Defaults(Boxplot) <- value}: Assign \code{Defaults} to a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Defaults(Kobe)}: Defaults \code{Code} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Defaults(Kobe) <- value}: Assign \code{Defaults} to a \code{\link[=Kobe-class]{Kobe-class()}} object

}}
