% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Design.R
\name{Design}
\alias{Design}
\alias{Design<-}
\title{Assign or access the Operating Model Design matrix}
\usage{
Design(Slick)

Design(Slick) <- value
}
\arguments{
\item{value}{A \code{data.frame} with the design matrix for the \code{OMs} object.
The columns must be named with the factor names. The rows contain the factor levels for each OM.
See \link{OMs} for more details.}

\item{object}{An object of class \code{\link[=Slick]{Slick()}} or class \code{\link[=OMs]{OMs()}}}
}
\description{
Assign or access the Operating Model Design matrix
}
\examples{
mySlick <- Slick()
Design(mySlick) <- data.frame(M=c(0.1, 0.2, 0.3),
                              h=c(0.6, 0.7, 0.8)
                             )
Design(mySlick)
}
