% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_OMs.R, R/class_Slick.R
\name{Factors}
\alias{Factors}
\alias{Factors<-}
\alias{Factors,OMs-method}
\alias{Factors<-,OMs-method}
\alias{Factors,Slick-method}
\alias{Factors<-,Slick-method}
\title{Return the Factors matrix from an \code{OMs} object}
\usage{
Factors(object, lang = "en")

Factors(object) <- value

\S4method{Factors}{OMs}(object, lang = "en")

\S4method{Factors}{OMs}(object) <- value

\S4method{Factors}{Slick}(object, lang = "en")

\S4method{Factors}{Slick}(object) <- value
}
\arguments{
\item{object}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{lang}{Optional text string specifying the language (if available).
Either 'en', 'es', 'fr', or 'pt' for English, Spanish, French, or Portuguese respectively}

\item{value}{A \code{data.frame} with the Factors}
}
\value{
The Design matrix from an \code{OMs} object in a \code{Slick} object
}
\description{
Return the Factors matrix from an \code{OMs} object
}
\section{Methods (by class)}{
\itemize{
\item \code{Factors(OMs)}: Return the operating model \code{Factors} table from a \code{\link[=OMs-class]{OMs-class()}} object

\item \code{Factors(OMs) <- value}: Assign the operating model \code{Factors} table  to a \code{\link[=OMs-class]{OMs-class()}} object

\item \code{Factors(Slick)}: Access the operating model \code{Factors} table from a \code{\link[=Slick-class]{Slick-class()}} object

\item \code{Factors(Slick) <- value}: Assign the operating model \code{Factors} table to a \code{\link[=Slick-class]{Slick-class()}} object

}}
