% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_casestudies.R
\name{get_casestudies}
\alias{get_casestudies}
\alias{download_casestudy}
\title{Get Case Studies from the SlickLibrary GitHub Repo}
\usage{
get_casestudies()

download_casestudy(
  name,
  case_studies = NULL,
  dir = NULL,
  silent = FALSE,
  object = TRUE,
  quiet = TRUE
)
}
\value{
A data.frame
}
\description{
Get Case Studies from the SlickLibrary GitHub Repo
}
\section{Functions}{
\itemize{
\item \code{download_casestudy()}: download a case study file

}}
