% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Kobe.R
\docType{class}
\name{Kobe-class}
\alias{Kobe-class}
\alias{show,Kobe-method}
\title{S4 class Kobe}
\usage{
\S4method{show}{Kobe}(object)
}
\arguments{
\item{object}{An object of class \code{\link[=Kobe-class]{Kobe-class()}}}
}
\description{
Slots can be accessed and assigned using functions corresponding to slot name.
See \verb{See Also} section below.
}
\details{
Objects of class \code{Kobe} are created with \code{Kobe()}
\subsection{Performance Indicators}{

There must be exactly two performance indicators (PIs).
The first PI will be on the x-axis (usually B/BMSY or something similar) and the second
on the y-axis (e.g., F/FMSY)
}

\subsection{Multi-Language Support}{

Text with multi-language supported can be provided as a named list. Available languages:
\itemize{
\item \code{en}: English (default)
\item \code{es}: Spanish
\item \code{fr}: French
}
}

\subsection{Note}{

Character strings in \code{Code}, \code{Label}, and \code{Description} must all be same length
as the number of performance indicators (\code{nPIs}) in \code{Value}
}
}
\section{Slots}{

\describe{
\item{\code{Code}}{A \emph{short} code for the Performance Indicators for this object.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{\code{Label}}{A short label for the Performance Indicators for this object. Used to label axes on charts.
Can be longer than \code{Code} but recommended to keep short as possible so it shows clearly in plots and tables.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{\code{Description}}{A description for the Performance Indicators for this object.
Can include Markdown, see \code{Examples}.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{\code{Time}}{A data.frame with a single column of values for the projection time-steps.
The column name will be used as the label in the \verb{Kobe Time} plot. Use a named list for
multiple languages.}

\item{\code{Value}}{A numeric array with the stochastic performance indicator values for each
simulation (sim), operating model (OM), management procedure (MP), performance indicator (PI),
and projection time-steps (nTS)
Dimensions: c(\code{nsim}, \code{nOM}, \code{nMP}, \code{nPI}, \code{nTS})}

\item{\code{Preset}}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. See \code{Examples}}
}}

\examples{
kobe <- Kobe(Code=c('PI1', 'PI2', 'PI3'),
                   Label=c('Performance Indicator 1',
                           'Performance Indicator 2',
                           'Performance Indicator 3'),
                   Description = c('This is the description for PI 1',
                                   'This is the description for PI 2',
                                   'This is the description for PI 3'),
                   Value=array(runif(3000), dim=c(10, 10, 10,3)),
                   Preset=list('PI and P2'=1:2)

)

kobe

# Code(boxplot)
# Metadata(boxplot)
#
# # Value(boxplot)
# Value(boxplot) <- array()
#
# Preset(boxplot)
# Preset(boxplot) <- list()
#
# boxplot
}
\seealso{
\code{\link[=Kobe-methods]{Kobe-methods()}}, \code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}},
\code{\link[=Metadata]{Metadata()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}}
}
