% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Metadata.R
\name{Metadata}
\alias{Metadata}
\title{Return Metadata Information}
\usage{
Metadata(object, lang = "en")
}
\arguments{
\item{object}{An object of class \code{\link[=MPs-class]{MPs-class()}}, \code{\link[=OMs-class]{OMs-class()}}, \code{\link[=Boxplot-class]{Boxplot-class()}},
\code{\link[=Kobe-class]{Kobe-class()}}, \code{\link[=Quilt-class]{Quilt-class()}}, \code{\link[=Spider-class]{Spider-class()}},
\code{\link[=Timeseries-class]{Timeseries-class()}}, or \code{\link[=Tradeoff-class]{Tradeoff-class()}}}

\item{lang}{Optional text string specifying the language (if available).
Either 'en', 'es', or 'fr' for English, Spanish, or French respectively}
}
\description{
Creates a data.frame with \code{Code}, \code{Label} and \code{Description} for objects of class
\code{MPs}, and all chart types. For objects of class \code{OMs}, it returns a data.frame with
\code{Factor}, \code{Level} and \code{Description}
}
