% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Quilt.R
\name{MinValue}
\alias{MinValue}
\alias{MinValue<-}
\alias{MaxValue}
\alias{MaxValue<-}
\title{Assign and access \code{MinValue} and \code{MaxValue} for a \code{Quilt} object}
\usage{
MinValue(Quilt)

MinValue(Quilt) <- value

MaxValue(Quilt)

MaxValue(Quilt) <- value
}
\arguments{
\item{Quilt}{A \code{\link[=Quilt-class]{Quilt-class()}} object}

\item{value}{A numeric vector with minimum or maximum values for the Performance Indicators.}
}
\description{
Assign and access \code{MinValue} and \code{MaxValue} for a \code{Quilt} object
}
\section{Functions}{
\itemize{
\item \code{MinValue()}: Return \code{MinValue} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{MinValue(Quilt) <- value}: Assign \code{MinValue} to a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{MaxValue()}: Return \code{MaxValue} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{MaxValue(Quilt) <- value}: Assign \code{MaxValue} to a \code{\link[=Quilt-class]{Quilt-class()}} object

}}
\seealso{
\code{\link[=Quilt]{Quilt()}}
}
