% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R
\name{Misc}
\alias{Misc}
\alias{Misc<-}
\title{Access or assign \code{Misc} for a valid object class}
\usage{
Misc(object)

Misc(object) <- value
}
\arguments{
\item{object}{A \code{\link[=Slick-class]{Slick-class()}}, \code{\link[=MPs-class]{MPs-class()}}, \code{\link[=Boxplot-class]{Boxplot-class()}}, \code{\link[=Kobe-class]{Kobe-class()}},
\code{\link[=Quilt-class]{Quilt-class()}}, \code{\link[=Spider-class]{Spider-class()}}, \code{\link[=Timeseries-class]{Timeseries-class()}}, or \code{\link[=Tradeoff-class]{Tradeoff-class()}} object}

\item{value}{A named list}
}
\description{
Access or assign \code{Misc} for a valid object class
}
