% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MPs.R, R/aa_generics.R
\docType{class}
\name{MPs-class}
\alias{MPs-class}
\alias{MPs,missingOrNULL-method}
\alias{MPs,dataframe_list-method}
\alias{MPs,Slick-method}
\alias{MPs<-,Slick-method}
\title{MPs S4 class and functions}
\usage{
MPs() # Create an empty MPs() object

MPs(Metadata=data.frame(), Preset=list()) # Populate a MPs() object

MPs(Slick, lang='en') # return MPs slot from Slick object

MPs(Slick) <- MPs() # assign an MPs object to the MPs slot in Slick object
}
\arguments{
\item{lang}{Optional. Character string length 2 to specify language (if available).  'en', 'es', or 'fr'}

\item{Slick}{An object of class \code{\link[=Slick]{Slick()}}}

\item{Metadata}{A \code{\link[=data.frame]{data.frame()}} with a specific structure. See \code{Details}}

\item{Preset}{An optional \code{\link[=list]{list()}} with a specific structure. See \code{Details}}
}
\description{
An object of class \code{MPs} contains information about the management procedures (MPs)
in the \code{\link[=Slick]{Slick()}} object.
}
\details{
\subsection{Metadata}{

The \code{Metadata} data.frame must have n MPs rows and the following columns:
\itemize{
\item Code: A short character string for identifying the MPs. Used in the \code{\link[=App]{App()}} in places
where there is no room for a full label.
\item Label: A long character string (one word or max two words) for identifying the MPs.
\item Description: A longer character string providing a description of the MPs.
\item Color: A character vector with the colors to use to display the MPs in the \code{\link[=App]{App()}}. Optional.
If not provided, the App will try to set some sensible colors.
}

For multi-language support, the \code{Metadata} slot can be populated with a named
list: names \code{en}, \code{es}, and \code{fr} for English, Spanish, and French respectively.
The data.frame within each list element must have identical structure; i.e., the same number of rows
and the column names in each list element must be identical.

The Metadata slot is accessed with \code{Metadata(MPs)} and modified with \code{Metadata(MPs) <- data.frame()}
}

\subsection{Preset}{

The \code{Preset} slot is an optional named list to add preset buttons for the Management
Procedure filters in the \code{\link[=App]{App()}}. The name of the list element will appear as a button
in the \code{\link[=App]{App()}}. Each list element should contain numeric values specifying the MPs to include.
The values must be <= n MPs.

The Preset slot is accessed with \code{Preset(MPs)} and modified with \verb{Preset(MPs) <- }list()`
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{MPs(missingOrNULL)}: Create a new object of class \code{\link[=MPs]{MPs()}}

\item \code{MPs(dataframe_list)}: Create a new object of class \code{\link[=MPs]{MPs()}}

\item \code{MPs(Slick)}: Return the \code{\link[=MPs]{MPs()}} object from an object of class \code{\link[=Slick]{Slick()}}

\item \code{MPs(Slick) <- value}: Assign an object of class \code{\link[=MPs]{MPs()}} to an object of class \code{\link[=Slick]{Slick()}}

}}
\section{Slots}{

\describe{
\item{\code{Metadata}}{A data.frame with a specific structure, describing the details of the management procedures. See \code{Details} section.}

\item{\code{Preset}}{An optional named list of preset buttons for filters in the \code{\link[=App]{App()}}. See \code{Details} section.}
}}

\examples{

myMPs <- MPs()
Metadata(myMPs) <- data.frame(Code=c('MP1', 'MP2', 'MP3'),
                              Label=c('Management Procedure 1',
                                      'Management Procedure 2',
                                      'Management Procedure 3'),
                              Description=c('This is the description for Management Procedure 1',
                                            'This is the description for Management Procedure 2',
                                            'This is the description for Management Procedure 3')
                              )
Preset(myMPs) <- list(All=1:3, FirstTwo=1:2)



# Multi-language
df_es <- Metadata(myMPs)
df_fr <- Metadata(myMPs)
df_es$Description <- c("This is the Spanish description for Management Procedure 1",
                       "This is the Spanish description for Management Procedure 2",
                       "This is the Spanish description for Management Procedure 3")
df_fr$Description <- c("This is the French description for Management Procedure 1",
                       "This is the French description for Management Procedure 2",
                       "This is the French description for Management Procedure 3")

Metadata(myMPs) <- list(en=Metadata(myMPs),
                        es=df_es,
                        fr=df_fr)


Metadata(myMPs)
Metadata(myMPs, 'es')
Metadata(myMPs, 'fr')
}
