% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OMs.R, R/aa_generics.R
\name{OMs}
\alias{OMs}
\alias{OMs,missing,missing,missing-method}
\alias{OMs,data.frame,missing,missing-method}
\alias{OMs,list,missing,missing-method}
\alias{OMs,data.frame,data.frame,missing-method}
\alias{OMs,data.frame,data.frame,list-method}
\alias{OMs,Slick,ANY,ANY-method}
\alias{OMs<-}
\alias{OMs<-,Slick-method}
\title{\code{OMs} S4 class and functions}
\usage{
OMs(Metadata, Design, Preset)

\S4method{OMs}{missing,missing,missing}()

\S4method{OMs}{data.frame,missing,missing}(Metadata)

\S4method{OMs}{list,missing,missing}(Metadata)

\S4method{OMs}{data.frame,data.frame,missing}(Metadata, Design)

\S4method{OMs}{data.frame,data.frame,list}(Metadata, Design, Preset)

\S4method{OMs}{Slick,ANY,ANY}(Metadata)

OMs(Slick) <- value

\S4method{OMs}{Slick}(Slick) <- value
}
\arguments{
\item{Metadata}{An object of class \code{\link[=data.frame]{data.frame()}} or \code{\link[=list]{list()}} formatted for the
\code{Metadata} slot. See \code{Details}. Or an object of class \link{Slick}.}

\item{Design}{A \code{data.frame} with the design matrix for the \code{OMs} object. See \code{Details}}

\item{Preset}{An optional named list. See \code{Details}}

\item{Slick}{An object of class \link{Slick}}

\item{value}{An object of class \link{OMs}}
}
\value{
An object of class \code{OMs}
}
\description{
An \code{OMs} object contains information about the operating models (MPs)
in the \code{\link[=Slick]{Slick()}} object.
}
\details{
TODO
\subsection{Metadata}{

TODO
}

\subsection{Design}{

TODO
}

\subsection{Preset}{

TODO
}
}
\section{Slots}{

\describe{
\item{\code{Metadata}}{TODO}

\item{\code{Design}}{TODO}

\item{\code{Preset}}{TODO}
}}

\examples{

myOMs <- OMs()

Metadata(myOMs) <- data.frame()


# Design <- expand.grid(M=c(0.1, 0.2, 0.3), h=c(0.7,0.9))
#' Description <- list(c('Natural mortality (M) = 0.1',
#'                       'Natural mortality (M) = 0.2',
#'                       'Natural mortality (M) = 0.3'),
#'                     c('Steepness (h) = 0.7',
#'                       'Steepness (h) = 0.9'))
#' Label <- list(c(M=0.1, M=0.2, M=0.3),
#'                c(h=0.7, h=0.9))
#' myOMs <- OMs(Design, Description, Label)

}
