% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotKobe.R
\name{plotKobe}
\alias{plotKobe}
\title{Plot \code{Kobe}}
\usage{
plotKobe(
  slick,
  xPI = 1,
  yPI = 2,
  TS = NA,
  Time = FALSE,
  BLcol = "#F8DC7A",
  TLcol = "#D8775D",
  TRcol = "#FDBD56",
  BRcol = "#67C18B",
  axis_label = "Code",
  percentile = 0.9,
  axis.text.size = 14,
  axis.title.size = 16,
  strip.text.size = 16,
  strip.text.color = "#D6501C",
  incMP_label = TRUE,
  mp.text.size = 7,
  mp.point.size = 4,
  mp.init.point.size = 2,
  xmax = 2,
  ymax = 2,
  hist_traj = FALSE,
  ncol = 4
)
}
\arguments{
\item{slick}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{xPI}{Numeric value specifying the performance indicator for the x-axis}

\item{yPI}{Numeric value specifying the performance indicator for the y-axis}

\item{TS}{Numeric value specifying the projection time-step for the Kobe plot. Default is NA
which means the terminal projection time-step}

\item{Time}{Logical. Kobe Time plot?}

\item{BLcol}{Color for the bottom left quadrant}

\item{TLcol}{Color for the top left quadrant}

\item{TRcol}{Color for the top  right quadrant}

\item{BRcol}{Color for the bottom right quadrant}

\item{axis_label}{Label to use for the axes. Either \code{Code} or \code{Label}.
\code{Description} works as well, but you probably don't want to do that.}

\item{axis.text.size}{Font size for axis text}

\item{axis.title.size}{Font size for axis title}

\item{strip.text.size}{Font size for facet strip text}

\item{strip.text.color}{Color for facet strip text}

\item{incMP_label}{Logical. Include MP labels?}

\item{mp.text.size}{Font size for MP labels}

\item{mp.point.size}{Point size for MP labels}

\item{mp.init.point.size}{Point size for start of trajectory. If \code{hist_traj==TRUE}}

\item{xmax}{Maximum value for the x-axis. Values greater than \code{xmax} will be shown at \code{xmax}}

\item{ymax}{Maximum value for the yx-axis. Values greater than \code{ymax} will be shown at \code{ymax}}

\item{hist_traj}{Logical. Plot the historical trajectories?}

\item{ncol}{Numeric. Number of columns for Kobe Time}

\item{percentile.}{Numeric value specifying the percentile for the x and y percentile bars.
Use NULL to remove percentile lines.}
}
\value{
A \code{ggplot2} object
}
\description{
Plots a Kobe plot for a given projection year, or a Kobe Time plot.
}
\seealso{
\code{\link[=Kobe-methods]{Kobe-methods()}}, \code{\link[=Kobe-class]{Kobe-class()}}
}
