% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotSpider.R
\name{plotSpider}
\alias{plotSpider}
\title{Plot \code{Spider}}
\usage{
plotSpider(
  slick,
  byOM = FALSE,
  byMP = FALSE,
  incMean = TRUE,
  incMax = TRUE,
  relScale = FALSE,
  size.pi.label = 6,
  size.mean = 4,
  size.mp.label = 16,
  size.max.value = 4,
  size.om.title = 5,
  col.om.title = "#D6501C",
  cex.om.title = 2,
  col.Mean = "white",
  bg.Mean = "black",
  grid.fill = "#f2f3f5",
  grid.line = "black",
  fill = byMP | all(byOM),
  inc.grid = TRUE,
  PI.labels = !(byMP | all(byOM)),
  MP_label = "Code",
  mp.lwd = 3,
  alpha = 0.3,
  ncol = 4,
  pm.avg.cex = 2.2,
  incMPtitle = byMP,
  mplab.cex = 2.2,
  max.pt.cex = 2,
  max.pt.col = "darkred"
)
}
\arguments{
\item{slick}{A \code{\link[=Slick-class]{Slick-class()}} object}
}
\value{
A \code{ggplot2} object
}
\description{
A Spider or Radar plot
}
\seealso{
\code{\link[=Slick-methods]{Slick-methods()}}, \code{\link[=Slick-class]{Slick-class()}}
}
