% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotTimeseries.R
\name{plotTimeseries}
\alias{plotTimeseries}
\title{Plot \code{Timeseries}}
\usage{
plotTimeseries(
  slick,
  PI = 1,
  byMP = FALSE,
  byOM = FALSE,
  includeHist = TRUE,
  ncol = 4,
  col_line = "darkgray",
  includeQuants = TRUE,
  includeLabels = TRUE,
  fill_ribbon1 = "#ededed",
  col_ribbon1 = "#ededed",
  quants1 = c(0.25, 0.75),
  alpha1 = 0.3,
  fill_ribbon2 = "white",
  col_ribbon2 = "#c9c9c9",
  linetype_ribbon2 = "dashed",
  quants2 = c(0.1, 0.9),
  alpha2 = 0.1,
  MP_label = "Code",
  size.title = 18,
  size.axis.title = 18,
  size.axis.text = 16,
  size.mp.label = 6,
  targ_color = "green",
  targ_name = "Target",
  lim_color = "red",
  lim_name = "Limit",
  inc_y_label = TRUE,
  sims = NULL
)
}
\arguments{
\item{slick}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{PI}{A numeric value specifying the performance indicator to plot}

\item{byMP}{Logical. Facet by MP? Defaults to FALSE, where all MPs are shown on the same plot}

\item{byOM}{Logical. Facet by OM?  Defaults to FALSE where values are calculated as mean across OMs}

\item{includeHist}{Logical. Include the historical period in the projections?}

\item{ncol}{Numeric. Number of columns if faceting by MP or OM}

\item{col_line}{Color for the median line (historical)}

\item{includeQuants}{Logical. Include quantile shading for the projections?}

\item{includeLabels}{Logical. Include MP labels?}

\item{fill_ribbon1}{Fill color for the inner ribbon}

\item{col_ribbon1}{Color of the line for inner ribbon}

\item{quants1}{Quantiles for the inner ribbon. Numeric length 2}

\item{alpha1}{Alpha for the colored MPs inner shading}

\item{fill_ribbon2}{Fill color for the outer ribbon}

\item{col_ribbon2}{Color of the line for outer ribbon}

\item{linetype_ribbon2}{Line type for outer ribbon}

\item{quants2}{Quantiles for the outer ribbon. Numeric length 2.}

\item{alpha2}{Alpha for the colored MPs outer shading}

\item{MP_label}{Label to use for the MPs. Either \code{Code} or \code{Label}.
\code{Description} works as well, but you probably don't want to do that.}

\item{size.title}{Numeric length 1. Size for plot title}

\item{size.axis.title}{Numeric length 1. Size for axis title}

\item{size.axis.text}{Numeric length 1. Size for axis text}

\item{size.mp.label}{Numeric length 1. Size of MP labels. Set to NULL for no MP labels}

\item{targ_color}{Color for the target line (if it exists in \code{Target(Timeseries(slick))})}

\item{targ_name}{Label for the target line}

\item{lim_color}{Color for the limit line (if it exists in \code{Limit(Timeseries(slick))})}

\item{lim_name}{Label for the limit line}

\item{inc_y_label}{Include the label for the y-axis?}

\item{sims}{Optional. Numeric values indicating the simulations to include. Defaults
to all.}
}
\value{
A \code{ggplot2} object
}
\description{
Plots the historical and projected values for a performance indicator.
}
\details{
If \code{byOM==FALSE} the results are shown as the mean across operating models.
}
\seealso{
\code{\link[=Timeseries-methods]{Timeseries-methods()}}, \code{\link[=Timeseries-class]{Timeseries-class()}}
}
