% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slick_tools.R
\name{PMnorm}
\alias{PMnorm}
\title{Normalize performance metric values in a Slick object}
\usage{
PMnorm(obj, det = T, inv = NULL)
}
\arguments{
\item{obj}{An object of class 'slick'}

\item{det}{logical, should the normalization be applied to the deterministic performance metrics (or, if false the stochastic ones)}

\item{inv}{A logical vector nPM long. If true, the PM will be inverted (100-value).}
}
\value{
An object of class \linkS4class{Slick}
}
\description{
A function that converts deterministic or stochastic performance metrics to the range 0-100 and optionally inverts these
}
\author{
T. Carruthers
}
