% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R
\name{Preset,Boxplot-method}
\alias{Preset,Boxplot-method}
\alias{Preset,Kobe-method}
\alias{Preset,Quilt-method}
\alias{Preset,Spider-method}
\alias{Preset,Timeseries-method}
\alias{Preset,Tradeoff-method}
\alias{Preset,MPs-method}
\alias{Preset<-,Boxplot-method}
\alias{Preset<-,Kobe-method}
\alias{Preset<-,Quilt-method}
\alias{Preset<-,Spider-method}
\alias{Preset<-,Timeseries-method}
\alias{Preset<-,Tradeoff-method}
\alias{Preset<-,MPs-method}
\title{Assign or access the \code{Preset} slot}
\usage{
\S4method{Preset}{Boxplot}(object)

\S4method{Preset}{Kobe}(object)

\S4method{Preset}{Quilt}(object)

\S4method{Preset}{Spider}(object)

\S4method{Preset}{Timeseries}(object)

\S4method{Preset}{Tradeoff}(object)

\S4method{Preset}{MPs}(object)

\S4method{Preset}{Boxplot}(object) <- value

\S4method{Preset}{Kobe}(object) <- value

\S4method{Preset}{Quilt}(object) <- value

\S4method{Preset}{Spider}(object) <- value

\S4method{Preset}{Timeseries}(object) <- value

\S4method{Preset}{Tradeoff}(object) <- value

\S4method{Preset}{MPs}(object) <- value
}
\arguments{
\item{object}{An object of class \link{Boxplot}, \link{Kobe}, \link{Quilt},
\link{Spider}, \link{Timeseries}, \link{Tradeoff}, or \link{MPs}}

\item{value}{A \code{list}, formatted to match the class of \code{object}. See the documentation for
corresponding \code{object} class for more details.}
}
\description{
Assign or access the \code{Preset} slot
}
