% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Quilt.R, R/class_Slick.R
\docType{methods}
\name{Quilt}
\alias{Quilt}
\alias{Quilt<-}
\alias{Quilt,missing-method}
\alias{Quilt,character_list-method}
\alias{Quilt,Slick-method}
\alias{Quilt<-,Slick-method}
\title{Methods for Creating, Accessing and Assigning \code{Quilt} objects}
\usage{
Quilt(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list(),
  Color = c("white", "darkblue"),
  MinValue = 0,
  MaxValue = as.numeric(NA)
)

Quilt(Slick) <- value

\S4method{Quilt}{missing}()

\S4method{Quilt}{character_list}(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list(),
  Color = c("white", "darkblue"),
  MinValue = 0,
  MaxValue = as.numeric(NA)
)

\S4method{Quilt}{Slick}(Code)

\S4method{Quilt}{Slick}(Slick) <- value
}
\arguments{
\item{Code}{A \emph{short} code for the Performance Indicators for this object.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Label}{A short label for the Performance Indicators for this object. Used to label axes on charts.
Can be longer than \code{Code} but recommended to keep short as possible so it shows clearly in plots and tables.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Description}{A description for the Performance Indicators for this object.
Can include Markdown, see \code{Examples}.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Value}{A numeric array with the stochastic performance indicator values for each
operating model (OM), management procedure (MP), and performance indicator (PI).
Dimensions: c(\code{nOM}, \code{nMP}, and \code{nPI})}

\item{Preset}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. The name of the list
element will appear as a button in the \code{\link[=App]{App()}}.}

\item{Color}{A character vector length 2 of colors for the maximum and minimum
values in the chart.}

\item{MinValue}{Numeric vector length \code{nPI} with the minimum possible value for the
respective PIs. Defaults to 0.}

\item{MaxValue}{Numeric vector length \code{nPI} with the maximum possible value (i.e., best performance)
for the respective PIs. Defaults to maximum PI value in \code{Value} (averaged across OMs in some cases).}

\item{Slick}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{value}{A \code{\link[=Quilt-class]{Quilt-class()}} object}
}
\description{
The \code{Quilt} function is used both to create and modify an \code{\link[=Quilt-class]{Quilt-class()}} object.
and to access and assign \code{Quilt} for an object of class \code{\link[=Slick-class]{Slick-class()}}.
See \code{Details}.
}
\details{
Objects of class \code{Quilt} are created with \code{Quilt()}

Use the  \code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}}, \code{\link[=Color]{Color()}},
\link{MinValue}, and \link{MaxValue} functions to access and assign the values for
an existing \code{Quilt} object, see \code{Examples}
\subsection{Multi-Language Support}{

Text with multi-language supported can be provided as a named list. Available languages:
\itemize{
\item \code{en}: English (default)
\item \code{es}: Spanish
\item \code{fr}: French
}
}

\subsection{Note}{

Character strings in \code{Code}, \code{Label}, and \code{Description} must all be same length
as the number of performance indicators (\code{nPIs}) in \code{Value}
}
}
\section{Functions}{
\itemize{
\item \code{Quilt(missing)}: Create an empty \code{Quilt} object

\item \code{Quilt(character_list)}: Create a populated \code{Quilt} object

\item \code{Quilt(Slick)}: Return \code{Quilt} from a \code{\link[=Slick-class]{Slick-class()}} object

\item \code{Quilt(Slick) <- value}: Assign a \code{\link[=Quilt-class]{Quilt-class()}} object to a \code{\link[=Slick-class]{Slick-class()}} object

}}
\seealso{
\code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Color]{Color()}}, \code{\link[=Metadata]{Metadata()}}, \code{\link[=Preset]{Preset()}},
\code{\link[=Color]{Color()}}, \code{\link[=MinValue]{MinValue()}}, \code{\link[=MaxValue]{MaxValue()}}
}
