% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Quilt.R, R/aa_generics.R
\name{Quilt}
\alias{Quilt}
\alias{Quilt,missing,missing,missing,missing-method}
\alias{Quilt,data.frame,ANY,ANY,ANY-method}
\alias{Quilt,list,ANY,ANY,ANY-method}
\alias{Quilt,data.frame,array,ANY,ANY-method}
\alias{Quilt,list,array,ANY,ANY-method}
\alias{Quilt,data.frame,array,list,missing-method}
\alias{Quilt,list,array,list,missing-method}
\alias{Quilt,data.frame,array,list,character-method}
\alias{Quilt,list,array,list,character-method}
\alias{Quilt,Slick,ANY,ANY,ANY-method}
\alias{Quilt<-,Slick-method}
\title{\code{Quilt} S4 class and functions}
\usage{
Quilt(Metadata, Value, Preset, Colors)

\S4method{Quilt}{missing,missing,missing,missing}()

\S4method{Quilt}{data.frame,ANY,ANY,ANY}(Metadata)

\S4method{Quilt}{list,ANY,ANY,ANY}(Metadata)

\S4method{Quilt}{data.frame,array,ANY,ANY}(Metadata, Value)

\S4method{Quilt}{list,array,ANY,ANY}(Metadata, Value)

\S4method{Quilt}{data.frame,array,list,missing}(Metadata, Value, Preset)

\S4method{Quilt}{list,array,list,missing}(Metadata, Value, Preset)

\S4method{Quilt}{data.frame,array,list,character}(Metadata, Value, Preset, Colors)

\S4method{Quilt}{list,array,list,character}(Metadata, Value, Preset, Colors)

\S4method{Quilt}{Slick,ANY,ANY,ANY}(Metadata)

\S4method{Quilt}{Slick}(Slick) <- value
}
\arguments{
\item{Metadata}{An object of class \code{\link[=data.frame]{data.frame()}} or \code{\link[=list]{list()}} formatted for the
\code{Metadata} slot. See \code{Details}. Or an object of class \link{Slick}.}

\item{Value}{An \code{array} with the values for the \code{Quilt} chart. See \code{Details}.
a character string specifying language (if available), or an object of class \code{\link[=MPs]{MPs()}}. See \code{Details}}

\item{Preset}{An optional named list. See \code{Details}.}

\item{Colors}{A character vector of length 2 specifying the colors for
the maximum and minimum values in the Quilt chart}

\item{Slick}{An object of class \link{Slick}}

\item{value}{An object of class \link{Quilt}}
}
\value{
An object of class \code{Quilt}
}
\description{
An object of class \code{Quilt} contains information for the Quilt chart.
The \code{Quilt} function is used both to create and modify an object of class \link{Quilt}
and to access and assign the \code{Quilt} slot in an object of class \link{Slick}. See \code{Details}.
}
\details{
TODO
\subsection{Metadata}{

TODO
}

\subsection{Value}{

TODO
}

\subsection{Preset}{

TODO
}

\subsection{Colors}{

TODO
}
}
\section{Slots}{

\describe{
\item{\code{Metadata}}{TODO}

\item{\code{Value}}{TODO}

\item{\code{Preset}}{TODO}

\item{\code{Color}}{TODO}
}}

