% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Timeseries.R
\name{RefPoints}
\alias{RefPoints}
\alias{RefPoints<-}
\alias{RefPoints,Timeseries-method}
\alias{RefPoints<-,Timeseries-method}
\title{Assign or access \code{RefPoints} for a valid object class}
\usage{
RefPoints(object)

RefPoints(object) <- value

\S4method{RefPoints}{Timeseries}(object)

\S4method{RefPoints}{Timeseries}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link[=Timeseries-class]{Timeseries-class()}}}

\item{value}{A \code{list}, formatted to match the class of \code{object}. See the documentation for
corresponding \code{object} class for more details.}
}
\value{
Returns a list object with the contents of the \code{RefPoints} slot
of \code{\link[=Timeseries]{Timeseries()}} objects
}
\description{
Assign or access \code{RefPoints} for a valid object class
}
\section{Methods (by class)}{
\itemize{
\item \code{RefPoints(Timeseries)}: Return \code{RefPoints} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{RefPoints(Timeseries) <- value}: Assign \code{RefPoints} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

}}
