% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R
\name{Selected,Tradeoff-method}
\alias{Selected,Tradeoff-method}
\alias{Selected<-,Tradeoff-method}
\title{Assign or access the \code{Selected} slot in an object of class \link{Tradeoff}}
\usage{
\S4method{Selected}{Tradeoff}(Tradeoff)

\S4method{Selected}{Tradeoff}(Tradeoff) <- value
}
\arguments{
\item{Tradeoff}{An object of class \link{Tradeoff}}

\item{value}{A character or numeric vector length 2. See \link{Tradeoff}}
}
\description{
Assign or access the \code{Selected} slot in an object of class \link{Tradeoff}
}
\examples{
tradeoff <- Tradeoff()
Selected(tradeoff) <- c(1,2)
Selected(tradeoff)
}
