% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Boxplot.R,
%   R/class_Check.R, R/class_Kobe.R, R/class_MPs.R, R/class_OMs.R,
%   R/class_Quilt.R, R/class_Tradeoff.R, R/class_Timeseries.R, R/class_Spider.R
\name{show}
\alias{show}
\alias{show,Boxplot-method}
\alias{show,CheckList-method}
\alias{show,Kobe-method}
\alias{show,MPs-method}
\alias{show,OMs-method}
\alias{show,Quilt-method}
\alias{show,Tradeoff-method}
\alias{show,Timeseries-method}
\alias{show,Spider-method}
\title{Generic show method}
\usage{
show(object)

\S4method{show}{Boxplot}(object)

\S4method{show}{CheckList}(object)

\S4method{show}{Kobe}(object)

\S4method{show}{MPs}(object)

\S4method{show}{OMs}(object)

\S4method{show}{Quilt}(object)

\S4method{show}{Tradeoff}(object)

\S4method{show}{Timeseries}(object)

\S4method{show}{Spider}(object)
}
\arguments{
\item{object}{Object to print to console}
}
\description{
Generic show method
}
\section{Methods (by class)}{
\itemize{
\item \code{show(Boxplot)}: Print a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{show(CheckList)}: Print a \code{CheckList} object

\item \code{show(Kobe)}: Print a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{show(MPs)}: Print a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{show(OMs)}: Print a \code{\link[=OMs-class]{OMs-class()}} object

\item \code{show(Quilt)}: Print a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{show(Tradeoff)}: Print a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{show(Timeseries)}: Print a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{show(Spider)}: Print a \code{\link[=Spider-class]{Spider-class()}} object

}}
