% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slick_functions.R
\name{Slick-functions}
\alias{Slick-functions}
\alias{Title}
\alias{Title<-}
\alias{Subtitle}
\alias{Subtitle<-}
\alias{Date}
\alias{Date<-}
\alias{Author}
\alias{Author<-}
\alias{Email}
\alias{Email<-}
\alias{Institution}
\alias{Institution<-}
\alias{Introduction}
\alias{Introduction<-}
\title{Slick Metadata Information}
\usage{
Title(Slick, lang = "es", markdown = TRUE)

Title(Slick) <- value

Subtitle(Slick, lang = "es", markdown = TRUE)

Subtitle(Slick) <- value

Date(Slick)

Date(Slick) <- value

Author(Slick, markdown = TRUE)

Author(Slick) <- value

Email(Slick, markdown = TRUE)

Email(Slick) <- value

Institution(Slick, markdown = TRUE)

Institution(Slick) <- value

Introduction(Slick, lang = "en", markdown = TRUE)

Introduction(Slick) <- value
}
\arguments{
\item{Slick}{An object of class \code{\link[=Slick-class]{Slick-class()}}}

\item{lang}{Optional text string specifying the language (if available).
Either 'en', 'es', or 'fr' for English, Spanish, or French respectively}

\item{markdown}{Logical. Process markdown?}

\item{value}{A value to assign to the \code{Slick} object. See \code{Details} and \code{Examples}}
}
\description{
Access or assign the metadata for a \code{\link[=Slick-class]{Slick-class()}} object
}
\details{
All functions with the exception of \code{Date} support Markdown.
\subsection{Multi-Langauge Support}{

\code{Title}, \code{Subtitle}, and \code{Introduction} have multi-language support.

Text with multi-language supported can be provided as a named list. Available languages:
\itemize{
\item \code{en}: English (default)
\item \code{es}: Spanish
\item \code{fr}: French
}

See \code{Examples}
}

\subsection{Assigning Values}{
\itemize{
\item \code{Title} A character string or named list with the title of the Slick object
\item \code{Subtitle} A character string or named list with the subtitle of the Slick object
\item \code{Date} Date the Slick object was created as text in format 'YYYY-MM-DD' or class \code{Date} e.g., \code{Sys.Date()}
\item \code{Author} A character string with Author(s) names. Should be length \code{nAuthors}
\item \code{Email} The email address corresponding with each author. Should be length \code{nAuthors}
\item \code{Institution} The institution corresponding with each author. Should be length \code{nAuthors}
\item \code{Introduction} Character string with introduction describing the Slick object. Length 1. See \code{Examples}
}
}
}
\section{Functions}{
\itemize{
\item \code{Title()}: Access \code{Title}, Multi-language support

\item \code{Title(Slick) <- value}: Assign \code{Title}, Multi-language support

\item \code{Subtitle()}: Access \code{Subtitle}, Multi-language support

\item \code{Subtitle(Slick) <- value}: Assign \code{Subtitle}, Multi-language support

\item \code{Date()}: Access \code{Date}

\item \code{Date(Slick) <- value}: Assign \code{Date}

\item \code{Author()}: Access \code{Author}

\item \code{Author(Slick) <- value}: Assign \code{Author}

\item \code{Email()}: Access \code{Email}

\item \code{Email(Slick) <- value}: Assign \code{Email}

\item \code{Institution()}: Access \code{Institution}

\item \code{Institution(Slick) <- value}: Assign \code{Institution}

\item \code{Introduction()}: Access \code{Introduction}

\item \code{Introduction(Slick) <- value}: Assign \code{Introduction}, can include Markdown. See \code{Examples}

}}
\examples{
# Assign values to a new `Slick` object
slick <- Slick()

Title(slick) <- 'An Example Slick Object'
Subtitle(slick) <- "[Created for the Developer's Guide](https://slick.bluematterscience.com/articles/Building-A-Slick-Object.Rmd)"
Date(slick) <- Sys.Date()
Author(slick) <- 'Adrian Hordyk'
Email(slick) <-  "[mailto:adrian@bluematterscience.com](mailto:adrian@bluematterscience.com)"
Institution(slick) <- "[Blue Matter Science](bluematterscience.com)"

Introduction(slick) <- "
This is an example Slick object. It has been created for the purposes of demonstrating the key features of Slick.

The introduction can include paragraphs as well as Markdown such as *italics* and **bold** text and [Links](https://slick.bluematterscience.com).
"

# Access values from `Slick` object
Title(slick)
Subtitle(slick)
Date(slick)
Author(slick)
Email(slick)
Institution(slick)
Introduction(slick)
}
