% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Slick.R
\docType{class}
\name{Slick-class}
\alias{Slick-class}
\alias{Slick}
\title{\code{Slick} S4 class and functions}
\usage{
Slick()
}
\description{
Each slot of the \code{Slick} object can be accessed or modified by the corresponding
accessor function. See Examples sections below for more details.
}
\section{Slots}{

\describe{
\item{\code{Title}}{Title for the \code{Slick} object. A character string. For multiple languages,
use a named list with names: \code{en}, \code{es}, \code{fr} for the three supported languages.}

\item{\code{Subtitle}}{Subtitle for the \code{Slick} object. A character string or a named list with
languages: \code{en}, \code{es}, \code{fr}}

\item{\code{Date}}{Either an object of class \code{Date} or character string.
Date the Slick object was created. Defaults to \code{Sys.Date()}}

\item{\code{Author}}{A character vector with name(s) of author(s) for the \code{Slick} object}

\item{\code{Email}}{A character vector with email addresses for the author(s). Optional.
Must be same length as \code{Author}}

\item{\code{Institution}}{A character vector with institution details for the author(s). Optional.
Must be same length as \code{Author}}

\item{\code{Introduction}}{Introduction text for the \code{Slick} object. Supports all markdown formatting.
For multiple languages, use a named list with names: \code{en}, \code{es}, \code{fr} for the
three supported languages.}

\item{\code{MPs}}{An object of class \code{\link[=MPs]{MPs()}}}

\item{\code{OMs}}{An object of class \code{\link[=OMs]{OMs()}}}

\item{\code{Boxplot}}{An object of class \code{\link[=Boxplot]{Boxplot()}}}

\item{\code{Kobe}}{An object of class \code{\link[=Kobe]{Kobe()}}}

\item{\code{Quilt}}{An object of class \code{\link[=Quilt]{Quilt()}}}

\item{\code{Spider}}{An object of class \code{\link[=Spider]{Spider()}}}

\item{\code{Timeseries}}{An object of class \code{\link[=Timeseries]{Timeseries()}}}

\item{\code{Tradeoff}}{An object of class \code{\link[=Tradeoff]{Tradeoff()}}}
}}

\examples{
mySlick <- Slick()
Title(mySlick) <- 'This is the title'
Subtitle(mySlick) <- 'This is the subtitle'
Date(mySlick) <- Sys.Date()

Title(mySlick)
Subtitle(mySlick)
Date(mySlick)


}
