% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slick2SlickData.R
\name{Slick2SlickData}
\alias{Slick2SlickData}
\title{Convert an object of class `Slick` to class `SlickData`}
\usage{
Slick2SlickData(slick)
}
\arguments{
\item{slick}{An object of class `Slick`}
}
\value{
An object of class `SlickData`
}
\description{
Previously objects loaded into Slick were class `Slick`. New class for these
objects is `SlickData`. This function converts the old class to the new class.
}
