% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SlickData.R
\docType{class}
\name{SlickData-class}
\alias{SlickData-class}
\alias{SlickData}
\title{Class \code{'SlickData'}}
\description{
A `Slick` Data object for uploading into the Slick App.
}
\author{
A. Hordyk
}
\keyword{classes}
