% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SlickData.R, R/aa_generics.R
\docType{class}
\name{SlickData-class}
\alias{SlickData-class}
\alias{SlickData}
\alias{Title,SlickData-method}
\alias{Title<-,SlickData-method}
\alias{Subtitle,SlickData-method}
\alias{Subtitle<-,SlickData-method}
\title{Class \code{'SlickData'}}
\usage{
SlickData() # creates an new `SlickData` object

SlickData(Title, Subtitle, Introduction)

\S4method{Title}{SlickData}(SlickData, lang = NULL)

\S4method{Title}{SlickData}(SlickData) <- value

\S4method{Subtitle}{SlickData}(SlickData, lang = NULL)

\S4method{Subtitle}{SlickData}(SlickData) <- value
}
\arguments{
\item{SlickData}{An object of class \code{SlickData}}

\item{lang}{two letter character string specifying the language
to return (if available). Currently supported languages are: \code{en}, \code{es}, \code{fr}}

\item{value}{Value to replace.}

\item{Title}{Title for the Slick object. A character string. For multiple languages,
use a named list with names: \code{en}, \code{es}, \code{fr} for the three supported languages.}

\item{Subtitle}{Subtitle for the Slick object. A character string or a named list with
languages: \code{en}, \code{es}, \code{fr}}

\item{Introduction}{Introduction text. A list where each list element is a
separate paragraph.}
}
\description{
A \code{Slick} Data object for uploading into the Slick App.
}
\author{
A. Hordyk
}
