% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Slick.R
\name{SlickFunctions}
\alias{SlickFunctions}
\alias{Title}
\alias{Title<-}
\alias{Subtitle}
\alias{Subtitle<-}
\alias{Date}
\alias{Date<-}
\alias{Author}
\alias{Author<-}
\alias{Email}
\alias{Email<-}
\alias{Institution}
\alias{Institution<-}
\alias{Introduction}
\alias{Introduction<-}
\title{Slick Object Functions}
\usage{
Title(object, lang = "en", markdown = FALSE)

Title(object) <- value

Subtitle(object, lang = "en", markdown = FALSE)

Subtitle(object) <- value

Date(object)

Date(object) <- value

Author(object, markdown = FALSE)

Author(object) <- value

Email(object, markdown = FALSE)

Email(object) <- value

Institution(object, lang = "en", markdown = FALSE)

Institution(object) <- value

Introduction(object, lang = "en", markdown = FALSE)

Introduction(object) <- value
}
\arguments{
\item{object}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{lang}{Optional text string specifying the language (if available).
Either 'en', 'es', or 'fr' for English, Spanish, or French respectively}

\item{markdown}{Logical. Process markdown?}

\item{value}{The value to assign to the object. See \code{Slots} for format of the
relevant object class}
}
\description{
Slick Object Functions
}
\section{Functions}{
\itemize{
\item \code{Title()}: Access \code{Title}, Multi-language support

\item \code{Title(object) <- value}: Assign \code{Title}, Multi-language support

\item \code{Subtitle()}: Access \code{Subtitle}, Multi-language support

\item \code{Subtitle(object) <- value}: Assign \code{Subtitle}, Multi-language support

\item \code{Date()}: Access \code{Date}

\item \code{Date(object) <- value}: Assign \code{Date}

\item \code{Author()}: Access \code{Author}

\item \code{Author(object) <- value}: Assign \code{Author}

\item \code{Email()}: Access \code{Email}

\item \code{Email(object) <- value}: Assign \code{Email}

\item \code{Institution()}: Access \code{Institution}

\item \code{Institution(object) <- value}: Assign \code{Institution}

\item \code{Introduction()}: Access \code{Introduction}

\item \code{Introduction(object) <- value}: Assign \code{Introduction}, can include Markdown. See \code{Examples}

}}
