% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Spider.R,
%   R/class_Slick.R
\docType{methods}
\name{Spider}
\alias{Spider}
\alias{Spider<-}
\alias{Spider,missing-method}
\alias{Spider,character-method}
\alias{Spider,list-method}
\alias{Spider,Slick-method}
\alias{Spider<-,Slick-method}
\title{Methods for Creating, Accessing and Assigning \code{Spider} objects}
\usage{
Spider(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list()
)

Spider(Slick) <- value

\S4method{Spider}{missing}()

\S4method{Spider}{character}(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list()
)

\S4method{Spider}{list}(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list()
)

\S4method{Spider}{Slick}(Code)

\S4method{Spider}{Slick}(Slick) <- value
}
\arguments{
\item{Code}{A \emph{short} code for the Performance Indicators for this object.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Label}{A short label for the Performance Indicators for this object. Used to label axes on charts.
Can be longer than \code{Code} but recommended to keep short as possible so it shows clearly in plots and tables.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Description}{A description for the Performance Indicators for this object.
Can include Markdown, see \code{Examples}.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Value}{A numeric array with the stochastic performance indicator values for each
operating model (OM), management procedure (MP), and performance indicator (PI).
Dimensions: c(\code{nOM}, \code{nMP}, and \code{nPI})}

\item{Preset}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. The name of the list
element will appear as a button in the \code{\link[=App]{App()}}.}

\item{Slick}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{value}{A \code{\link[=Spider-class]{Spider-class()}} object}
}
\description{
The \code{Spider} function is used both to create and modify an \code{\link[=Spider-class]{Spider-class()}} object.
and to access and assign \code{Spider} for an object of class \code{\link[=Slick-class]{Slick-class()}}.
See \code{Details}.
}
\details{
Objects of class \code{Spider} are created with \code{Spider()}

Use the  \code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}} functions to
access and assign the values for an existing \code{Spider} object, see \code{Examples}
\subsection{Multi-Language Support}{

Text with multi-language supported can be provided as a named list. Available languages:
\itemize{
\item \code{en}: English (default)
\item \code{es}: Spanish
\item \code{fr}: French
}
}

\subsection{Note}{

Character strings in \code{Code}, \code{Label}, and \code{Description} must all be same length
as the number of performance indicators (\code{nPIs}) in `Value
}
}
\section{Functions}{
\itemize{
\item \code{Spider(missing)}: Create an empty \code{Spider} object

\item \code{Spider(character)}: Create a populated \code{Spider} object

\item \code{Spider(list)}: Create a populated \code{Spider} object

\item \code{Spider(Slick)}: Return \code{Spider} from a \code{\link[=Slick-class]{Slick-class()}} object

\item \code{Spider(Slick) <- value}: Assign a \code{\link[=Spider-class]{Spider-class()}} object to a \code{\link[=Slick-class]{Slick-class()}} object

}}
\seealso{
\code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Metadata]{Metadata()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}}
}
