% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Spider.R, R/aa_generics.R
\name{Spider}
\alias{Spider}
\alias{Spider,missing,missing,missing-method}
\alias{Spider,data.frame,missing,missing-method}
\alias{Spider,list,missing,missing-method}
\alias{Spider,list,array,missing-method}
\alias{Spider,list,array,list-method}
\alias{Spider,Slick,ANY,ANY-method}
\alias{Spider<-,Slick-method}
\title{\code{Spider} S4 class and functions}
\usage{
Spider(Metadata, Value, Preset)

\S4method{Spider}{missing,missing,missing}()

\S4method{Spider}{data.frame,missing,missing}(Metadata)

\S4method{Spider}{list,missing,missing}(Metadata)

\S4method{Spider}{list,array,missing}(Metadata, Value)

\S4method{Spider}{list,array,list}(Metadata, Value, Preset)

\S4method{Spider}{Slick,ANY,ANY}(Metadata)

\S4method{Spider}{Slick}(Slick) <- value
}
\arguments{
\item{Metadata}{An object of class \code{\link[=data.frame]{data.frame()}} or \code{\link[=list]{list()}} formatted for the
\code{Metadata} slot. See \code{Details}. Or an object of class \link{Slick}.}

\item{Preset}{An optional named list. See \code{Details}.}

\item{Slick}{An object of class \link{Slick}}

\item{value}{An object of class \link{Spider}}
}
\value{
An object of class \code{Spider}
}
\description{
An object of class \code{Spider} contains information for the Spider chart.
The \code{Spider} function is used both to create and modify an object of class \link{Spider}
and to access and assign the \code{Spider} slot in an object of class \link{Slick}. See \code{Details}.
}
\details{
TODO
\subsection{Metadata}{

TODO
}

\subsection{Value}{

TODO
}

\subsection{Preset}{

TODO
}
}
\section{Slots}{

\describe{
\item{\code{Metadata}}{TODO}

\item{\code{Value}}{TODO}

\item{\code{Preset}}{TODO}
}}

