% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R
\name{Table,Boxplot,character,ANY-method}
\alias{Table,Boxplot,character,ANY-method}
\alias{Table,Boxplot,missing,ANY-method}
\alias{Table,Kobe,character,ANY-method}
\alias{Table,Kobe,missing,ANY-method}
\alias{Table,Quilt,character,ANY-method}
\alias{Table,Quilt,missing,ANY-method}
\alias{Table,Spider,character,ANY-method}
\alias{Table,Spider,missing,ANY-method}
\alias{Table,Timeseries,character,ANY-method}
\alias{Table,Timeseries,missing,ANY-method}
\alias{Table,Tradeoff,character,ANY-method}
\alias{Table,Tradeoff,missing,ANY-method}
\alias{Table,MPs,character,ANY-method}
\alias{Table,MPs,missing,ANY-method}
\alias{Table,OMs,character,ANY-method}
\alias{Table,OMs,missing,ANY-method}
\alias{Table,OMs,character,character-method}
\title{Create a DT::datatable table from information in Metadata slot}
\usage{
\S4method{Table}{Boxplot,character,ANY}(object, lang)

\S4method{Table}{Boxplot,missing,ANY}(object)

\S4method{Table}{Kobe,character,ANY}(object, lang)

\S4method{Table}{Kobe,missing,ANY}(object)

\S4method{Table}{Quilt,character,ANY}(object, lang)

\S4method{Table}{Quilt,missing,ANY}(object)

\S4method{Table}{Spider,character,ANY}(object, lang)

\S4method{Table}{Spider,missing,ANY}(object)

\S4method{Table}{Timeseries,character,ANY}(object, lang)

\S4method{Table}{Timeseries,missing,ANY}(object)

\S4method{Table}{Tradeoff,character,ANY}(object, lang)

\S4method{Table}{Tradeoff,missing,ANY}(object)

\S4method{Table}{MPs,character,ANY}(object, lang)

\S4method{Table}{MPs,missing,ANY}(object)

\S4method{Table}{OMs,character,ANY}(object, lang)

\S4method{Table}{OMs,missing,ANY}(object)

\S4method{Table}{OMs,character,character}(object, lang, type)
}
\arguments{
\item{object}{An object of class \link{Boxplot}, \link{Kobe}, \link{Quilt},
\link{Spider}, \link{Timeseries}, \link{Tradeoff}, or \link{MPs} or \link{OMs}}

\item{lang}{Optional text string specifying the language (if available).
Either 'en', 'es', or 'fr' for English, Spanish, or French respectively}

\item{type}{For objects of class \link{OMs}, create a table for the 'factor' or 'design' slot?}
}
\description{
Create a \link[DT:datatable]{DT::datatable} table from information in \link{Metadata} slot.
}
