% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Kobe.R,
%   R/class_Timeseries.R
\name{Limit}
\alias{Limit}
\alias{Limit<-}
\alias{Target}
\alias{Target<-}
\alias{Limit,Kobe-method}
\alias{Limit<-,Kobe-method}
\alias{Target,Kobe-method}
\alias{Target<-,Kobe-method}
\alias{Limit,Timeseries-method}
\alias{Limit<-,Timeseries-method}
\alias{Target,Timeseries-method}
\alias{Target<-,Timeseries-method}
\title{Access or assign \code{Target} and \code{Limit} for object of class \code{Kobe} or \code{Timeseries}}
\usage{
Limit(object)

Limit(object) <- value

Target(object)

Target(object) <- value

\S4method{Limit}{Kobe}(object)

\S4method{Limit}{Kobe}(object) <- value

\S4method{Target}{Kobe}(object)

\S4method{Target}{Kobe}(object) <- value

\S4method{Limit}{Timeseries}(object)

\S4method{Limit}{Timeseries}(object) <- value

\S4method{Target}{Timeseries}(object)

\S4method{Target}{Timeseries}(object) <- value
}
\arguments{
\item{object}{A \code{\link[=Kobe-class]{Kobe-class()}} or \code{\link[=Timeseries]{Timeseries()}} class object}

\item{value}{Value to assign to \code{Target}}
}
\value{
Returns a numeric vector with the contents of the \code{Target} or \code{Limit}
slots of \code{\link[=Kobe]{Kobe()}} and
\code{\link[=Timeseries]{Timeseries()}} objects
}
\description{
Access or assign \code{Target} and \code{Limit} for object of class \code{Kobe} or \code{Timeseries}
}
\section{Methods (by class)}{
\itemize{
\item \code{Limit(Kobe)}: Return \code{Limit} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Limit(Kobe) <- value}: Assign \code{Limit} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Target(Kobe)}: Return \code{Target} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Target(Kobe) <- value}: Assign \code{Target} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Limit(Timeseries)}: Return \code{Limit} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Limit(Timeseries) <- value}: Assign \code{Limit} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Target(Timeseries)}: Return \code{Target} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Target(Timeseries) <- value}: Assign \code{Target} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

}}
\section{Functions}{
\itemize{
\item \code{Target(object) <- value}: Assign \code{value} to \code{object@Target}

}}
