% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R
\name{Time,Kobe-method}
\alias{Time,Kobe-method}
\alias{Time<-,Kobe-method}
\alias{Time<-,Timeseries-method}
\title{Access or assign the Time slot in object of class \code{Kobe} or \code{Timeseries}}
\usage{
\S4method{Time}{Kobe}(object)

\S4method{Time}{Kobe}(object) <- value

\S4method{Time}{Timeseries}(object) <- value
}
\arguments{
\item{value}{A list or data.frame formatted for the class of \code{object}}
}
\description{
Access or assign the Time slot in object of class \link{Kobe} or \link{Timeseries}
}
