% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Kobe.R,
%   R/class_Timeseries.R
\name{TimeLab}
\alias{TimeLab}
\alias{TimeLab<-}
\alias{TimeLab,Kobe-method}
\alias{TimeLab<-,Kobe-method}
\alias{TimeLab,Timeseries-method}
\alias{TimeLab<-,Timeseries-method}
\title{Access or assign \code{TimeLab} in a \code{Kobe} or \code{Timeseries} object}
\usage{
TimeLab(object, lang = "en")

TimeLab(object) <- value

\S4method{TimeLab}{Kobe}(object, lang = "en")

\S4method{TimeLab}{Kobe}(object) <- value

\S4method{TimeLab}{Timeseries}(object, lang = "en")

\S4method{TimeLab}{Timeseries}(object) <- value
}
\arguments{
\item{object}{A \code{\link[=Kobe-class]{Kobe-class()}} or \linkS4class{Timeseries} object}

\item{lang}{Optional text string specifying the language (if available).
Either 'en', 'es', 'fr', or 'pt' for English, Spanish, French, or Portuguese respectively}

\item{value}{A character string to assign to \code{TimeLab} in \code{object}.}
}
\value{
Returns a character string from the \code{TimeLab} slot in
\code{\link[=Kobe]{Kobe()}} and \code{\link[=Timeseries]{Timeseries()}} objects
}
\description{
Access or assign \code{TimeLab} in a \code{Kobe} or \code{Timeseries} object
}
\section{Methods (by class)}{
\itemize{
\item \code{TimeLab(Kobe)}: Return \code{TimeLab} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{TimeLab(Kobe) <- value}: Assign \code{TimeLab} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{TimeLab(Timeseries)}: Return \code{TimeLab} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{TimeLab(Timeseries) <- value}: Assign \code{TimeLab} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

}}
