% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Timeseries.R
\name{TimeNow}
\alias{TimeNow}
\alias{TimeNow<-}
\title{Access and return \code{TimeNow} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object}
\usage{
TimeNow(Timeseries)

TimeNow(Timeseries) <- value
}
\arguments{
\item{Timeseries}{A \code{\link[=Timeseries-class]{Timeseries-class()}} object}

\item{value}{A character string label for the time axis. Use a named list for
multiple languages}
}
\description{
Access and return \code{TimeNow} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object
}
\section{Functions}{
\itemize{
\item \code{TimeNow(Timeseries) <- value}: Assign \code{TimeNow} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

}}
