% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Timeseries.R
\name{Timeseries}
\alias{Timeseries}
\title{Create a new \code{Timeseries} object}
\usage{
Timeseries(x, ...)
}
\value{
An object of class \code{Timeseries}
}
\description{
Create a new \code{Timeseries} object
}
