% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Kobe.R
\name{TimeTerminal}
\alias{TimeTerminal}
\alias{TimeTerminal<-}
\alias{TimeTerminal,Kobe-method}
\alias{TimeTerminal<-,Kobe-method}
\title{Assign or access \code{TimeTerminal} for a valid object class}
\usage{
TimeTerminal(object)

TimeTerminal(object) <- value

\S4method{TimeTerminal}{Kobe}(object)

\S4method{TimeTerminal}{Kobe}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link[=Kobe-class]{Kobe-class()}}}

\item{value}{A \code{numeric} value with a value matching one in \code{Kobe@Time} \code{\link[=Kobe-class]{Kobe-class()}} for details.}
}
\value{
Returns a numeric values from the \code{TimeTerminal} slot in \code{\link[=Kobe]{Kobe()}}
objects
}
\description{
Assign or access \code{TimeTerminal} for a valid object class
}
\section{Methods (by class)}{
\itemize{
\item \code{TimeTerminal(Kobe)}: Return \code{TimeTerminal} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{TimeTerminal(Kobe) <- value}: Assign \code{TimeTerminal} to a \code{\link[=Kobe-class]{Kobe-class()}} object

}}
