% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tradeoff.R, R/aa_generics.R
\name{Tradeoff}
\alias{Tradeoff}
\alias{Tradeoff,missing,missing,missing,missing-method}
\alias{Tradeoff,data.frame,missing,missing,missing-method}
\alias{Tradeoff,data.frame,array,missing,missing-method}
\alias{Tradeoff,data.frame,array,list,missing-method}
\alias{Tradeoff,data.frame,array,list,character-method}
\alias{Tradeoff,data.frame,array,list,numeric-method}
\alias{Tradeoff,list,missing,missing,missing-method}
\alias{Tradeoff,list,array,missing,missing-method}
\alias{Tradeoff,list,array,list,missing-method}
\alias{Tradeoff,list,array,list,character-method}
\alias{Tradeoff,list,array,list,numeric-method}
\alias{Tradeoff,Slick,ANY,ANY,ANY-method}
\alias{Tradeoff<-,Slick-method}
\title{\code{Tradeoff} S4 class and functions}
\usage{
Tradeoff(Metadata, Value, Preset, Selected)

\S4method{Tradeoff}{missing,missing,missing,missing}()

\S4method{Tradeoff}{data.frame,missing,missing,missing}(Metadata)

\S4method{Tradeoff}{data.frame,array,missing,missing}(Metadata, Value)

\S4method{Tradeoff}{data.frame,array,list,missing}(Metadata, Value, Preset)

\S4method{Tradeoff}{data.frame,array,list,character}(Metadata, Value, Preset, Selected)

\S4method{Tradeoff}{data.frame,array,list,numeric}(Metadata, Value, Preset, Selected)

\S4method{Tradeoff}{list,missing,missing,missing}(Metadata)

\S4method{Tradeoff}{list,array,missing,missing}(Metadata, Value)

\S4method{Tradeoff}{list,array,list,missing}(Metadata, Value, Preset)

\S4method{Tradeoff}{list,array,list,character}(Metadata, Value, Preset, Selected)

\S4method{Tradeoff}{list,array,list,numeric}(Metadata, Value, Preset, Selected)

\S4method{Tradeoff}{Slick,ANY,ANY,ANY}(Metadata)

\S4method{Tradeoff}{Slick}(Slick) <- value
}
\arguments{
\item{Metadata}{An object of class \code{\link[=data.frame]{data.frame()}} or \code{\link[=list]{list()}} formatted for the
\code{Metadata} slot. See \code{Details}. Or an object of class \link{Slick}.}

\item{Value}{An \code{array} with the values for the \code{Boxplot} chart. See \code{Details}.}

\item{Preset}{An optional named list. See \code{Details}.}

\item{Selected}{TODO}

\item{An}{object of class \link{Tradeoff}}
}
\value{
An object of class \code{Tradeoff}
}
\description{
An object of class \code{Tradeoff} contains information for the Time Series chart.
The \code{Tradeoff} function is used both to create and modify an object of class \link{Tradeoff}
and to access and assign the \code{Tradeoff} slot in an object of class \link{Slick}. See \code{Details}.
}
\details{
\subsection{Creating objects of class \link{Tradeoff}}{
\itemize{
\item \code{Tradeoff()} creates a new object of class \link{Tradeoff}
}
}

\subsection{Accessing and assigning \code{Tradeoff} slot in objects of class \link{Slick}}{

TODO
}

\subsection{Metadata}{

TODO
}

\subsection{Value}{

TODO
}

\subsection{Preset}{

TODO
}

\subsection{Selected}{
}
}
\section{Slots}{

\describe{
\item{\code{Metadata}}{TODO}

\item{\code{Value}}{TODO}

\item{\code{Preset}}{TODO}

\item{\code{Selected}}{TODO}
}}

