% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R
\name{Value,Boxplot-method}
\alias{Value,Boxplot-method}
\alias{Value,Kobe-method}
\alias{Value,Quilt-method}
\alias{Value,Spider-method}
\alias{Value,Timeseries-method}
\alias{Value,Tradeoff-method}
\alias{Value<-,Boxplot-method}
\alias{Value<-,Kobe-method}
\alias{Value<-,Quilt-method}
\alias{Value<-,Spider-method}
\alias{Value<-,Timeseries-method}
\alias{Value<-,Tradeoff-method}
\title{Assign or access the \code{Value} slot}
\usage{
\S4method{Value}{Boxplot}(object)

\S4method{Value}{Kobe}(object)

\S4method{Value}{Quilt}(object)

\S4method{Value}{Spider}(object)

\S4method{Value}{Timeseries}(object)

\S4method{Value}{Tradeoff}(object)

\S4method{Value}{Boxplot}(object) <- value

\S4method{Value}{Kobe}(object) <- value

\S4method{Value}{Quilt}(object) <- value

\S4method{Value}{Spider}(object) <- value

\S4method{Value}{Timeseries}(object) <- value

\S4method{Value}{Tradeoff}(object) <- value
}
\arguments{
\item{object}{An object of class \link{Boxplot}, \link{Kobe}, \link{Quilt},
\link{Spider}, \link{Timeseries}, or \link{Tradeoff}}

\item{value}{An \link{array}, formatted to match the class of \code{object}. See the documentation for
corresponding \code{object} class for more details.}
}
\description{
Assign or access the \code{Value} slot
}
