% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSlick.R
\name{App}
\alias{App}
\title{Run the Slick App in the local browser}
\usage{
App(...)
}
\arguments{
\item{...}{Additional arguments. To load a object directly in the App,
use \code{App(slick=myslick)} where \code{myslick} is your \code{\link[=Slick-class]{Slick-class()}} object.}
}
\value{
None
}
\description{
An online hosted version of the Slick App is available
\href{https://shiny.bluematterscience.com/app/slick}{here}
}
\details{
\code{App()} runs the Slick Shiny Application
}
