% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Boxplot.R,
%   R/class_Kobe.R, R/class_MPs.R, R/class_Quilt.R, R/class_Tradeoff.R,
%   R/class_Timeseries.R, R/class_Spider.R
\name{Code}
\alias{Code}
\alias{Code<-}
\alias{Description}
\alias{Description<-}
\alias{Label}
\alias{Label<-}
\alias{Code,Boxplot-method}
\alias{Code<-,Boxplot-method}
\alias{Description,Boxplot-method}
\alias{Label,Boxplot-method}
\alias{Label<-,Boxplot-method}
\alias{Description<-,Boxplot-method}
\alias{Misc,Boxplot-method}
\alias{Misc<-,Boxplot-method}
\alias{Code,Kobe-method}
\alias{Code<-,Kobe-method}
\alias{Description,Kobe-method}
\alias{Description<-,Kobe-method}
\alias{Label,Kobe-method}
\alias{Label<-,Kobe-method}
\alias{Misc,Kobe-method}
\alias{Misc<-,Kobe-method}
\alias{Code,MPs-method}
\alias{Code<-,MPs-method}
\alias{Label,MPs-method}
\alias{Label<-,MPs-method}
\alias{Description,MPs-method}
\alias{Description<-,MPs-method}
\alias{Code,Quilt-method}
\alias{Code<-,Quilt-method}
\alias{Description,Quilt-method}
\alias{Description<-,Quilt-method}
\alias{Label,Quilt-method}
\alias{Label<-,Quilt-method}
\alias{Misc,Quilt-method}
\alias{Misc<-,Quilt-method}
\alias{Code,Tradeoff-method}
\alias{Code<-,Tradeoff-method}
\alias{Description,Tradeoff-method}
\alias{Description<-,Tradeoff-method}
\alias{Misc,Tradeoff-method}
\alias{Misc<-,Tradeoff-method}
\alias{Label,Tradeoff-method}
\alias{Label<-,Tradeoff-method}
\alias{Code,Timeseries-method}
\alias{Code<-,Timeseries-method}
\alias{Description,Timeseries-method}
\alias{Description<-,Timeseries-method}
\alias{Label,Timeseries-method}
\alias{Label<-,Timeseries-method}
\alias{Misc,Timeseries-method}
\alias{Misc<-,Timeseries-method}
\alias{Code,Spider-method}
\alias{Code<-,Spider-method}
\alias{Description,Spider-method}
\alias{Description<-,Spider-method}
\alias{Label,Spider-method}
\alias{Label<-,Spider-method}
\alias{Misc,Spider-method}
\alias{Misc<-,Spider-method}
\title{Access or assign \code{Code}, \code{Label}, and \code{Description} for a valid object class}
\usage{
Code(object, lang = "en")

Code(object) <- value

Description(object, lang = "en")

Description(object) <- value

Label(object, lang = "en")

Label(object) <- value

\S4method{Code}{Boxplot}(object, lang = "en")

\S4method{Code}{Boxplot}(object) <- value

\S4method{Description}{Boxplot}(object, lang = "en")

\S4method{Label}{Boxplot}(object, lang = "en")

\S4method{Label}{Boxplot}(object) <- value

\S4method{Description}{Boxplot}(object) <- value

\S4method{Misc}{Boxplot}(object)

\S4method{Misc}{Boxplot}(object) <- value

\S4method{Code}{Kobe}(object, lang = "en")

\S4method{Code}{Kobe}(object) <- value

\S4method{Description}{Kobe}(object, lang = "en")

\S4method{Description}{Kobe}(object) <- value

\S4method{Label}{Kobe}(object, lang = "en")

\S4method{Label}{Kobe}(object) <- value

\S4method{Misc}{Kobe}(object)

\S4method{Misc}{Kobe}(object) <- value

\S4method{Code}{MPs}(object, lang = "en")

\S4method{Code}{MPs}(object) <- value

\S4method{Label}{MPs}(object, lang = "en")

\S4method{Label}{MPs}(object) <- value

\S4method{Description}{MPs}(object, lang = "en")

\S4method{Description}{MPs}(object) <- value

\S4method{Code}{Quilt}(object, lang = "en")

\S4method{Code}{Quilt}(object) <- value

\S4method{Description}{Quilt}(object, lang = "en")

\S4method{Description}{Quilt}(object) <- value

\S4method{Label}{Quilt}(object, lang = "en")

\S4method{Label}{Quilt}(object) <- value

\S4method{Misc}{Quilt}(object)

\S4method{Misc}{Quilt}(object) <- value

\S4method{Code}{Tradeoff}(object, lang = "en")

\S4method{Code}{Tradeoff}(object) <- value

\S4method{Description}{Tradeoff}(object, lang = "en")

\S4method{Description}{Tradeoff}(object) <- value

\S4method{Misc}{Tradeoff}(object)

\S4method{Misc}{Tradeoff}(object) <- value

\S4method{Label}{Tradeoff}(object, lang = "en")

\S4method{Label}{Tradeoff}(object) <- value

\S4method{Code}{Timeseries}(object, lang = "en")

\S4method{Code}{Timeseries}(object) <- value

\S4method{Description}{Timeseries}(object, lang = "en")

\S4method{Description}{Timeseries}(object) <- value

\S4method{Label}{Timeseries}(object, lang = "en")

\S4method{Label}{Timeseries}(object) <- value

\S4method{Misc}{Timeseries}(object)

\S4method{Misc}{Timeseries}(object) <- value

\S4method{Code}{Spider}(object, lang = "en")

\S4method{Code}{Spider}(object) <- value

\S4method{Description}{Spider}(object, lang = "en")

\S4method{Description}{Spider}(object) <- value

\S4method{Label}{Spider}(object, lang = "en")

\S4method{Label}{Spider}(object) <- value

\S4method{Misc}{Spider}(object)

\S4method{Misc}{Spider}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link[=MPs-class]{MPs-class()}}, \code{\link[=Boxplot-class]{Boxplot-class()}},
\code{\link[=Kobe-class]{Kobe-class()}}, \code{\link[=Quilt-class]{Quilt-class()}}, \code{\link[=Spider-class]{Spider-class()}},
\code{\link[=Timeseries-class]{Timeseries-class()}}, or \code{\link[=Tradeoff-class]{Tradeoff-class()}}}

\item{lang}{Optional text string specifying the language (if available).
Either 'en', 'es', or 'fr' for English, Spanish, or French respectively}

\item{value}{A character vector or a named list for multi-language support}
}
\value{
Returns character string or named list with the contents of the \code{Code},
\code{Label}, or \code{Description} slot of \code{object}
}
\description{
Access or assign \code{Code}, \code{Label}, and \code{Description} for a valid object class
}
\details{
\code{Code}, \code{Label}, and \code{Description} must all be equal length.
\subsection{Multi-Language Support}{

Text with multi-language supported can be provided as a named list. Available languages:
\itemize{
\item \code{en}: English (default)
\item \code{es}: Spanish
\item \code{fr}: French
}
}
}
\section{Methods (by class)}{
\itemize{
\item \code{Code(Boxplot)}: Return \code{Code} from a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Code(Boxplot) <- value}: Assign \code{Code} to a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Description(Boxplot)}: Return \code{Description} from a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Label(Boxplot)}: Return \code{Label} from a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Label(Boxplot) <- value}: Assign \code{Label} to a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Description(Boxplot) <- value}: Assign \code{Description} to a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Misc(Boxplot)}: Return \code{Misc} from a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Misc(Boxplot) <- value}: Assign \code{Misc} to a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Code(Kobe)}: Return \code{Code} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Code(Kobe) <- value}: Assign \code{Code} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Description(Kobe)}: Return \code{Description} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Description(Kobe) <- value}: Assign \code{Description} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Label(Kobe)}: Return \code{Label} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Label(Kobe) <- value}: Assign \code{Label} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Misc(Kobe)}: Return \code{Misc} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Misc(Kobe) <- value}: Assign \code{Misc} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Code(MPs)}: Return \code{Code} from a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Code(MPs) <- value}: Assign \code{Code} to a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Label(MPs)}: Return \code{Label} from a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Label(MPs) <- value}: Assign \code{Label} to a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Description(MPs)}: Return \code{Description} from a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Description(MPs) <- value}: Assign \code{Description} to a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Code(Quilt)}: Return \code{Code} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Code(Quilt) <- value}: Assign \code{Code} to a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Description(Quilt)}: Return \code{Description} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Description(Quilt) <- value}: Assign \code{Description} to a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Label(Quilt)}: Return \code{Label} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Label(Quilt) <- value}: Assign \code{Label} to a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Misc(Quilt)}: Return \code{Misc} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Misc(Quilt) <- value}: Assign \code{Misc} to a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Code(Tradeoff)}: Return \code{Code} from a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Code(Tradeoff) <- value}: Assign \code{Code} to a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Description(Tradeoff)}: Return \code{Description} from a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Description(Tradeoff) <- value}: Assign \code{Description} to a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Misc(Tradeoff)}: Return \code{Misc} from a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Misc(Tradeoff) <- value}: Assign \code{Misc} to a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Label(Tradeoff)}: Return \code{Label} from a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Label(Tradeoff) <- value}: Assign \code{Label} to a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Code(Timeseries)}: Return \code{Code} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Code(Timeseries) <- value}: Assign \code{Code} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Description(Timeseries)}: Return \code{Description} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Description(Timeseries) <- value}: Assign \code{Description} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Label(Timeseries)}: Return \code{Label} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Label(Timeseries) <- value}: Assign \code{Label} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Misc(Timeseries)}: Return \code{Misc} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Misc(Timeseries) <- value}: Assign \code{Misc} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Code(Spider)}: Return \code{Code} from a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Code(Spider) <- value}: Assign \code{Code} to a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Description(Spider)}: Return \code{Description} from a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Description(Spider) <- value}: Assign \code{Description} to a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Label(Spider)}: Return \code{Label} from a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Label(Spider) <- value}: Assign \code{Label} to a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Misc(Spider)}: Return \code{Misc} from a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Misc(Spider) <- value}: Assign \code{Misc} to a \code{\link[=Spider-class]{Spider-class()}} object

}}
\seealso{
\code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=MPs-class]{MPs-class()}}, \code{\link[=Boxplot-class]{Boxplot-class()}},
\code{\link[=Kobe-class]{Kobe-class()}}, \code{\link[=Quilt-class]{Quilt-class()}}, \code{\link[=Spider-class]{Spider-class()}},
\code{\link[=Timeseries-class]{Timeseries-class()}}, \code{\link[=Tradeoff-class]{Tradeoff-class()}}
}
