% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_OMs.R
\name{Design}
\alias{Design}
\alias{Design<-}
\alias{Design,OMs-method}
\alias{Design<-,OMs-method}
\title{Return the Design matrix from an \code{OMs} object}
\usage{
Design(object)

Design(object) <- value

\S4method{Design}{OMs}(object)

\S4method{Design}{OMs}(object) <- value
}
\arguments{
\item{object}{An \code{\link[=OMs-class]{OMs-class()}} or \code{\link[=Slick-class]{Slick-class()}} object.}

\item{value}{A \code{data.frame} with the Design matrix}
}
\value{
The Design matrix from an \code{OMs} object in a \code{Slick} object
}
\description{
Return the Design matrix from an \code{OMs} object
}
\section{Methods (by class)}{
\itemize{
\item \code{Design(OMs)}: Return the operating model \code{Design} matrix from a \code{\link[=OMs-class]{OMs-class()}} object

\item \code{Design(OMs) <- value}: Assign the operating model \code{Design} matrix to a \code{\link[=OMs-class]{OMs-class()}} object

}}
