% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Kobe.R, R/class_Slick.R
\docType{methods}
\name{Kobe}
\alias{Kobe}
\alias{Kobe<-}
\title{Methods for Creating, Accessing and Assigning \code{Kobe} objects}
\usage{
Kobe(Code, Label, Description, Time, Value, Preset)

Kobe(Slick) <- value
}
\arguments{
\item{Code}{A \emph{short} code for the Performance Indicators for this object.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Label}{A short label for the Performance Indicators for this object. Used to label axes on charts.
Can be longer than \code{Code} but recommended to keep short as possible so it shows clearly in plots and tables.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Description}{A description for the Performance Indicators for this object.
Can include Markdown, see \code{Examples}.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Time}{A data.frame with a single column of values for the projection time-steps.
The column name will be used as the label in the \verb{Kobe Time} plot. Use a named list for
multiple languages.}

\item{Value}{A numeric array with the stochastic performance indicator values for each
simulation (sim), operating model (OM), management procedure (MP), performance indicator (PI),
and projection time-steps (nTS).
Dimensions: c(\code{nsim}, \code{nOM}, \code{nMP}, \code{nPI}, \code{nTS})}

\item{Preset}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. See \code{Examples}}

\item{Slick}{An object of class \code{\link[=Slick-class]{Slick-class()}}}

\item{value}{An object of class \code{\link[=Kobe-class]{Kobe-class()}}}
}
\description{
An object of class \code{Kobe} contains information for the Kobe chart.
The \code{Kobe} function is used both to create and modify an \code{\link[=Kobe-class]{Kobe-class()}} object.
and to access and assign \code{Kobe} for an object of class \code{\link[=Slick-class]{Slick-class()}}.
See \code{Details}.
}
\details{
The Kobe plot typically shows B/BMSY (or something similar) on the x-axis, and
F/FMSY (or something similar) on the y-axis.
\subsection{Performance Indicators}{

There must be exactly two performance indicators (PIs).
The first PI will be on the x-axis (usually B/BMSY or something similar) and the second
on the y-axis (e.g., F/FMSY)
}

\subsection{Multi-Language Support}{

Text with multi-language supported can be provided as a named list. Available languages:
\itemize{
\item \code{en}: English (default)
\item \code{es}: Spanish
\item \code{fr}: French
}
}

\subsection{Note}{

Character strings in \code{Code}, \code{Label}, and \code{Description} must all be same length
as the number of performance indicators (\code{nPIs}) in \code{Value}
\itemize{
\item \code{Kobe()}: creates an empty \code{Kobe} object
\item \code{Kobe(Code, ...)}: creates a populated \code{Kobe} object
\item \code{Kobe(Slick)}: returns the \code{Kobe} from a \code{Slick} object
\item \code{Kobe(Slick) <- Kobe}: assigns a \code{Kobe} object to a \code{Slick} object
}

Use the  \code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}} functions to access and assign the values for an
existing \code{Kobe} object, see \code{Examples}
}
}
\examples{
kobe <- Kobe(Code=c('PI1', 'PI2', 'PI3'),
                   Label=c('Performance Indicator 1',
                           'Performance Indicator 2',
                           'Performance Indicator 3'),
                   Description = c('This is the description for PI 1',
                                   'This is the description for PI 2',
                                   'This is the description for PI 3'),
                   Value=array(runif(3000), dim=c(10, 10, 10,3)),
                   Preset=list('PI and P2'=1:2)

)

kobe

# Code(boxplot)
# Metadata(boxplot)
#
# # Value(boxplot)
# Value(boxplot) <- array()
#
# Preset(boxplot)
# Preset(boxplot) <- list()
#
# boxplot
}
\seealso{
\code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Metadata]{Metadata()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}}
}
