% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R
\name{Metadata}
\alias{Metadata}
\alias{Metadata,MPs-method}
\alias{Metadata<-,MPs-method}
\alias{Metadata,OMs-method}
\alias{Metadata<-,OMs-method}
\title{Access or modify the \code{Metadata} slot}
\usage{
Metadata(x, ...)

\S4method{Metadata}{MPs}(x, lang = "en")

\S4method{Metadata}{MPs}(x) <- value

\S4method{Metadata}{OMs}(x, lang = "en")

\S4method{Metadata}{OMs}(x) <- value
}
\arguments{
\item{x}{An object with a \code{Metadata} slot}

\item{...}{Additional arguments, specific to the class of object \code{x}}

\item{lang}{Character string to select language.
Either 'en', 'es', or 'fr' for English, Spanish, or French respectively.
If selected language isn't available in the object, it defaults to first language}

\item{value}{A correctly structured data.frame for object class \code{x}}
}
\description{
Access or modify the \code{Metadata} slot
}
\section{Methods (by class)}{
\itemize{
\item \code{Metadata(MPs)}: Modify the \code{Metadata} slot in an \code{\link[=MPs]{MPs()}} object

\item \code{Metadata(MPs) <- value}: Assign the \code{Metadata} slot in an \code{\link[=MPs]{MPs()}} object

\item \code{Metadata(OMs)}: Modify the \code{Metadata} slot in an \code{\link[=OMs]{OMs()}} object

\item \code{Metadata(OMs) <- value}: Assign the \code{Metadata} slot in an \code{\link[=OMs]{OMs()}} object

}}
