% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MPs.R
\docType{class}
\name{MPs-class}
\alias{MPs-class}
<<<<<<< Updated upstream
\alias{MPs}
\alias{MPs,dataframe_list-method}
\alias{MPs,missingOrNULL-method}
\alias{MPs,Slick-method}
\alias{MPs<-}
\alias{MPs<-,Slick-method}
\title{MPs S4 class and functions}
\usage{
MPs(x, ...)

\S4method{MPs}{dataframe_list}(x, ...)

\S4method{MPs}{missingOrNULL}(x, ...)

\S4method{MPs}{Slick}(x, lang = NULL)

MPs(Slick) <- value

\S4method{MPs}{Slick}(Slick) <- value
}
\arguments{
\item{x}{An object of class \code{\link[=data.frame]{data.frame()}} or \code{\link[=list]{list()}}}

\item{...}{additional arguments (not used)}

\item{lang}{optional.character string length 2 to specify language (if available).  'en', 'es', or 'fr'}

\item{Slick}{An object of class \code{\link[=Slick]{Slick()}}}

\item{value}{An object of class \code{\link[=MPs]{MPs()}}}

\item{missingOrNULL}{missing or NULL}
}
\description{
An \code{MPs} object contains information about the management procedures (MPs)
in the \code{\link[=Slick]{Slick()}} object.
=======
\title{\code{MPs} S4 class and functions}
\description{
An object of class \code{MPs} contains information about the management procedures (MPs)
in the \code{\link[=Slick]{Slick()}} object. The \code{MPs} function is used both to create and modify an object of class \code{\link[=MPs]{MPs()}} and to access
and assign the \code{MPs} slot in an object of class \code{\link[=Slick]{Slick()}}. See \code{Details}.
>>>>>>> Stashed changes
}
\details{
\subsection{Creating objects of class \code{\link[=MPs]{MPs()}}}{
\itemize{
\item \code{MPs()} creates a new object of class \code{\link[=MPs]{MPs()}}
\item \code{MPs(object)} creates an \code{MPs} object with \code{object} assigned to the \code{\link[=Metadata]{Metadata()}} slot,
where \code{object} is either a \code{data.frame} or a named \code{list}.  See \code{Metadata} section below.
\item \code{MPs(object, value)} creates an \code{MPs} object with \code{object} assigned to the \link{Metadata}
slot and \code{value} assigned to the \code{Preset} slot,  where \code{object} is as described above,
and \code{value} is a named \code{list}.  See \code{Preset} section below.
}
}

\subsection{Accessing and assigning slots in objects of class \link{MPs}}{
\itemize{
\item The \code{Metadata} slot is accessed with \code{Metadata(MPs)} and modified with \code{Metadata(MPs) <- value}
\item The \code{Preset} slot is accessed with \code{Preset(MPs)} and modified with \code{Preset(MPs) <- value}
}

See \code{\link[=Metadata]{Metadata()}} and \code{\link[=Preset]{Preset()}} for more details.
}

\subsection{Accessing and assigning \code{MPs} slot in objects of class \code{\link[=Slick]{Slick()}}}{
\itemize{
\item \code{MPs(object)} accesses the \code{MPs} slot in an object of class \code{\link[=Slick]{Slick()}}
\item \code{MPs(object) <- value} assigns the \code{MPs} slot in an object of class \code{\link[=Slick]{Slick()}},
where \code{value} is an object of class \code{\link[=MPs]{MPs()}}
}
\subsection{Metadata}{

The \code{Metadata} data.frame must have n MPs rows and the following columns:
\itemize{
\item Code: A short character string for identifying the MPs. Used in the \code{\link[=App]{App()}} in places
where there is no room for a full label.
\item Label: A long character string (one word or max two words) for identifying the MPs.
\item Description: A longer character string providing a description of the MPs.
\item Color: A character vector with the colors to use to display the MPs in the \code{\link[=App]{App()}}. Optional.
If not provided, the App will try to set some sensible colors.
}

For multi-language support, the \code{Metadata} slot can be populated with a named
list: names \code{en}, \code{es}, and \code{fr} for English, Spanish, and French respectively.
The data.frame within each list element must have identical structure; i.e., the same number of rows
and the column names in each list element must be identical.
}

\subsection{Preset}{

The \code{Preset} slot is an optional named list to add preset buttons for the Management
Procedure filters in the \code{\link[=App]{App()}}. The name of the list element will appear as a button
in the \code{\link[=App]{App()}}. Each list element should contain numeric values specifying the MPs to include.
The values must be <= n MPs.

The Preset slot is accessed with \code{Preset(MPs)} and modified with \verb{Preset(MPs) <- }list()`
}

}
}
<<<<<<< Updated upstream
\section{Methods (by generic)}{
\itemize{
\item \code{MPs(dataframe_list)}: Create a new object of class \code{\link[=MPs]{MPs()}}

\item \code{MPs(missingOrNULL)}: Create a new object of class \code{\link[=MPs]{MPs()}}

\item \code{MPs(Slick)}: Return the \code{\link[=MPs]{MPs()}} object from an object of class \code{\link[=Slick]{Slick()}}

\item \code{MPs(Slick) <- value}: Assign an object of class \code{\link[=MPs]{MPs()}} to an object of class \code{\link[=Slick]{Slick()}}

}}
\section{Functions}{
\itemize{
\item \code{MPs()}: A generic function

\item \code{MPs(Slick) <- value}: Assign an object of class \code{\link[=MPs]{MPs()}} to an object of class \code{\link[=Slick]{Slick()}}

}}
=======
>>>>>>> Stashed changes
\section{Slots}{

\describe{
\item{\code{Metadata}}{A data.frame with a specific structure, describing the details of the management procedures. See \code{Details} section.}

\item{\code{Preset}}{An optional named list of preset buttons for filters in the \code{\link[=App]{App()}}. See \code{Details} section.}
}}

\examples{

myMPs <- MPs()
Metadata(myMPs) <- data.frame(Code=c('MP1', 'MP2', 'MP3'),
                              Label=c('Management Procedure 1',
                                      'Management Procedure 2',
                                      'Management Procedure 3'),
                              Description=c('This is the description for Management Procedure 1',
                                            'This is the description for Management Procedure 2',
                                            'This is the description for Management Procedure 3')
                              )
Preset(myMPs) <- list(All=1:3, FirstTwo=1:2)



# Multi-language
df_es <- Metadata(myMPs)
df_fr <- Metadata(myMPs)
df_es$Description <- c("This is the Spanish description for Management Procedure 1",
                       "This is the Spanish description for Management Procedure 2",
                       "This is the Spanish description for Management Procedure 3")
df_fr$Description <- c("This is the French description for Management Procedure 1",
                       "This is the French description for Management Procedure 2",
                       "This is the French description for Management Procedure 3")

Metadata(myMPs) <- list(en=Metadata(myMPs),
                        es=df_es,
                        fr=df_fr)


Metadata(myMPs)
Metadata(myMPs, 'es')
Metadata(myMPs, 'fr')
}
