% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OMs.R
\docType{class}
\name{OMs-class}
\alias{OMs-class}
\alias{show,OMs-method}
\title{\code{OMs} S4 class and functions}
\usage{
\S4method{show}{OMs}(object)
}
\description{
An object of class \code{OMs} contains information about the operating models (MPs)
in the \code{\link[=Slick]{Slick()}} object. Like all S4 objects in \code{Slick}, slots in this
object can be accessed and assigned using functions corresponding to slot name.
See \code{\link[=OMs]{OMs()}} and the the \verb{See Also} section below.
}
\details{
\subsection{Multi-Language Support}{

Use a named list to use multi-languages in \code{Factors}
}

\subsection{Factors}{

\code{Factors} can be accessed and assigned using \code{Factors(myslick)} and
\code{Factors(myslick) <- data.frame()} respectively.

The \code{Factor} column should be character strings with the name of each factor,
while the \code{Level} column is a \code{numeric} or \code{character} value with the level for the
corresponding factor.

The \code{Description} column is a description for each row, i.e., a unique factor and level.
See \code{Examples}.
}

\subsection{Design}{

The \code{Design} matrix is \code{nOM} rows and \code{nFactor} columns. The values in each column should
either be \code{numeric} values indicating the levels for the corresponding factor,
or the actual level values (i.e., \code{Factors$Level}) that correspond to each OM. See \code{Examples}.
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(OMs)}: Show objects of class \code{OMs}

}}
\section{Slots}{

\describe{
\item{\code{Factors}}{A \code{data.frame} with column headings \code{Factor}, \code{Level}, and \code{Description}. See \code{Details}}

\item{\code{Design}}{A \code{data.frame} with \code{nFactor} columns (i.e., \code{length(unique(Factors$Factor))}), and \code{nOM}
rows. See \code{Details}}

\item{\code{Preset}}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. The name of the list
element will appear as a button in the \code{\link[=App]{App()}}.}
}}

\examples{

myOMs <- OMs()

Metadata(myOMs) <- data.frame()


# Design <- expand.grid(M=c(0.1, 0.2, 0.3), h=c(0.7,0.9))
#' Description <- list(c('Natural mortality (M) = 0.1',
#'                       'Natural mortality (M) = 0.2',
#'                       'Natural mortality (M) = 0.3'),
#'                     c('Steepness (h) = 0.7',
#'                       'Steepness (h) = 0.9'))
#' Label <- list(c(M=0.1, M=0.2, M=0.3),
#'                c(h=0.7, h=0.9))
#' myOMs <- OMs(Design, Description, Label)

#' mySlick <- Slick()
#' Design(mySlick) <- data.frame(M=c(0.1, 0.2, 0.3),
#'                               h=c(0.6, 0.7, 0.8)
#'                              )
#' Design(mySlick)
}
\seealso{
\code{\link[=OMs-methods]{OMs-methods()}}, \code{\link[=Factors]{Factors()}}, \code{\link[=Design]{Design()}}, \code{\link[=Preset]{Preset()}}
}
