% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_OMs.R, R/class_Slick.R
\docType{methods}
\name{OMs}
\alias{OMs}
\alias{OMs<-}
\alias{OMs,missing-method}
\alias{OMs,dataframe_list-method}
\alias{OMs,Slick-method}
\alias{OMs<-,Slick-method}
\title{Methods for Creating, Accessing and Assigning \code{OMs} objects}
\usage{
OMs(Factors = data.frame(), Design = data.frame(), Preset = list())

OMs(object) <- value

\S4method{OMs}{missing}()

\S4method{OMs}{dataframe_list}(Factors = data.frame(), Design = data.frame(), Preset = list())

\S4method{OMs}{Slick}(Factors)

\S4method{OMs}{Slick}(object) <- value
}
\arguments{
\item{Factors}{A \code{data.frame} with column headings \code{Factor}, \code{Level}, and \code{Description}.
See \code{Details}}

\item{Design}{A \code{data.frame} with \code{nFactor} columns
(i.e., \code{length(unique(Factors$Factor))}), and \code{nOM}
rows. See \code{Details}}

\item{Preset}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. The name of the list
element will appear as a button in the \code{\link[=App]{App()}}.}

\item{object}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{value}{A \code{\link[=OMs-class]{OMs-class()}} object}
}
\description{
The \code{OMs} function is used both to create and modify an \code{\link[=OMs-class]{OMs-class()}} object.
and to access and assign \code{OMs} for an object of class \code{\link[=Slick-class]{Slick-class()}}.
See \code{Details}.
}
\details{
\subsection{Factors}{

\code{Factors} can be accessed and assigned using \code{Factors(myslick)} and
\code{Factors(myslick) <- data.frame()} respectively.

The \code{Factor} column should be character strings with the name of each factor,
while the \code{Level} column is a \code{numeric} or \code{character} value with the level for the
corresponding factor.

The \code{Description} column is a description for each row, i.e., a unique factor and level.
See \code{Examples}.
}

\subsection{Design}{

The \code{Design} matrix is \code{nOM} rows and \code{nFactor} columns. The values in each column should
either be \code{numeric} values indicating the levels for the corresponding factor,
or the actual level values (i.e., \code{Factors$Level}) that correspond to each OM. See \code{Examples}.

Use \code{\link[=Factors]{Factors()}}, \code{\link[=Design]{Design()}}, and \code{\link[=Preset]{Preset()}} to access
and assign the values for an existing \code{OMs} object, see \code{Examples}.
}
}
\section{Functions}{
\itemize{
\item \code{OMs(missing)}: Create an empty \code{OMs} object

\item \code{OMs(dataframe_list)}: Create a populated \code{OMs} object

\item \code{OMs(Slick)}: Return an \code{\link[=OMs-class]{OMs-class()}} object from a \code{\link[=Slick]{Slick()}} object

\item \code{OMs(Slick) <- value}: Assign an \code{\link[=OMs-class]{OMs-class()}} object to a \code{\link[=Slick]{Slick()}} object

}}
\examples{

myOMs <- OMs()

Metadata(myOMs) <- data.frame()


# Design <- expand.grid(M=c(0.1, 0.2, 0.3), h=c(0.7,0.9))
#' Description <- list(c('Natural mortality (M) = 0.1',
#'                       'Natural mortality (M) = 0.2',
#'                       'Natural mortality (M) = 0.3'),
#'                     c('Steepness (h) = 0.7',
#'                       'Steepness (h) = 0.9'))
#' Label <- list(c(M=0.1, M=0.2, M=0.3),
#'                c(h=0.7, h=0.9))
#' myOMs <- OMs(Design, Description, Label)

#' mySlick <- Slick()
#' Design(mySlick) <- data.frame(M=c(0.1, 0.2, 0.3),
#'                               h=c(0.6, 0.7, 0.8)
#'                              )
#' Design(mySlick)
}
\seealso{
\code{\link[=OMs-class]{OMs-class()}}, \code{\link[=Factors]{Factors()}}, \code{\link[=Design]{Design()}}, \code{\link[=Preset]{Preset()}}
}
