% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OMs.R
\name{OMs}
\alias{OMs}
\title{Create or modify an \code{OMs} object}
\usage{
OMs()
}
\arguments{
\item{Metadata}{A data.frame with a specific structure, describing the details of
the operating models. See \code{Details} section}

\item{Design}{A \code{data.frame} with the design matrix for the OMs. The columns
must be named with the factor names. The rows contain the factor levels for each OM. See \code{Details} section}

\item{Preset}{An optional named list of preset buttons for filters in the \code{\link[=App]{App()}}. See \code{Details} section.}
}
\value{
An object of class \code{OMs}
}
\description{
An \code{OMs} object contains information about the operating models (MPs)
in the \code{\link[=Slick]{Slick()}} object.
}
\details{
\code{OMs()} creates an empty \code{OMs} object.

Each slot of the \code{OMs} object can be accessed or modified by the corresponding
accessor function. See \code{Examples} sections below for more details.
\subsection{Metadata}{

The \code{Metadata} data.frame must have n OMs rows and the following columns:
\itemize{
\item Code: A short character string for identifying the MPs. Used in the \code{\link[=App]{App()}} in places
where there is no room for a full label.
\item Label: A long character string (one word or max two words) for identifying the MPs.
\item Description: A longer character string providing a description of the MPs.
\item Color: A character vector with the colors to use to display the MPs in the \code{\link[=App]{App()}}. Optional.
If not provided, the App will try to set some sensible colors.
}
}

\subsection{Design}{
}

\subsection{Preset}{
}
}
\examples{

myOMs <- OMs()

Metadata(myOMs) <- data.frame()


# Design <- expand.grid(M=c(0.1, 0.2, 0.3), h=c(0.7,0.9))
#' Description <- list(c('Natural mortality (M) = 0.1',
#'                       'Natural mortality (M) = 0.2',
#'                       'Natural mortality (M) = 0.3'),
#'                     c('Steepness (h) = 0.7',
#'                       'Steepness (h) = 0.9'))
#' Label <- list(c(M=0.1, M=0.2, M=0.3),
#'                c(h=0.7, h=0.9))
#' myOMs <- OMs(Design, Description, Label)
}
