% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OMs.R, R/aa_generics.R
\name{OMs}
\alias{OMs}
\alias{OMs<-}
\alias{OMs,data.frame-method}
\alias{OMs,missingOrNULL-method}
\alias{OMs,SlickData-method}
\alias{OMs<-,SlickData-method}
\title{Generic function for \code{OMs} objects}
\usage{
OMs(Design, Description, Label, Default)

OMs(x, ...)

OMs(x) <- value

\S4method{OMs}{data.frame}(x, ...)

\S4method{OMs}{missingOrNULL}(x, ...)

\S4method{OMs}{SlickData}(x)

\S4method{OMs}{SlickData}(x) <- value
}
\arguments{
\item{Design}{A \code{data.frame} with the design matrix for the OMs. The columns
must be named with the factor names. The rows contain the factor levels for each OM.}

\item{Description}{A list of length n factors (i.e., \code{ncol(Design)}), with each
element a character vector of length n levels (i.e., \code{lapply(lapply(Design, unique), length)}),
with a description of the factor levels.}

\item{Label}{(optional) A list with the same structure as \code{Description},
with a short label to be used in the plots. If missing, the label will be the
factor level.}

\item{Default}{(optional) A \code{numeric} vector specifying the rows in \code{Design}
to be included when Slick is first run.
If not specified, all OMs will be selected when Slick starts.}
}
\value{
An object of class \code{OMs}
}
\description{
Creates or populates a new \code{OMs} object, or returns or assigns the \code{OMs} slot in objects of class \code{SlickData}
}
\details{
Creates an empty \code{OMs} object, or specifies the design matrix, description, and labels of the operating models
(OMs) included in the Slick Data object.

Also used to return or assign the \code{OMs} slot for objects of class \code{SlickData}
}
\examples{
Design <- expand.grid(M=c(0.1, 0.2, 0.3), h=c(0.7,0.9))
Description <- list(c('Natural mortality (M) = 0.1',
                      'Natural mortality (M) = 0.2',
                      'Natural mortality (M) = 0.3'),
                    c('Steepness (h) = 0.7',
                      'Steepness (h) = 0.9'))
Label <- list(c(M=0.1, M=0.2, M=0.3),
               c(h=0.7, h=0.9))
myOMs <- OMs(Design, Description, Label)
}
