% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotBoxplot.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Plot \code{Boxplot}}
\usage{
plotBoxplot(
  slick,
  PI = NULL,
  type = c("boxplot", "violin", "both", "all"),
  byOM = FALSE,
  ncol = 4,
  MP_label = "Code",
  PI_label = "Code"
)
}
\arguments{
\item{slick}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{PI}{Numeric value indicating the Performance Indicator(s) to plot from the \code{Boxplot-class} object.
If NULL, it will facet_wrap all PIs}

\item{type}{Character string specifying the plot type.}

\item{byOM}{Logical. Facet the plots by operating model? PI must be a single value}

\item{ncol}{Numeric. Number of columns}

\item{MP_label}{Label to use for the MPs. Either \code{Code} or \code{Label}.
\code{Description} works as well, but you probably don't want to do that.}

\item{PI_label}{Label to use for the PIs. Either \code{Code} or \code{Label}.
\code{Description} works as well, but you probably don't want to do that.}
}
\value{
A \code{ggplot2} object, or a list of \code{ggplot2} objects
}
\description{
Plots boxplot, violin plot, or a combined box+violin plot for information stored
in a \linkS4class{Boxplot} object
}
\examples{
boxplot <- Boxplot(Code=c('PI1', 'PI2', 'PI3'),
                   Label=c('Performance Indicator 1',
                         'Performance Indicator 2',
                         'Performance Indicator 3'),
                   Description = c('This is the description for PI 1',
                                   'This is the description for PI 2',
                                   'This is the description for PI 3'),
                   Value=array(runif(3000), dim=c(6, 5, 4, 3)),
                   Preset=list('PI and P2'=1:2)

)

boxplot

Code(boxplot) <- c('LTY', 'STY', 'P50')
Metadata(boxplot)


# Add to `Slick` object
slick <- Slick()
Boxplot(slick) <- boxplot

plotBoxplot(slick)

plotBoxplot(slick, type='violin')

plotBoxplot(slick, type='both')
}
