% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Plot \code{Boxplot}}
\usage{
plotBoxplot(
  slick,
  pm = 1,
  type = c("boxplot", "violin", "both", "all"),
  byOM = FALSE,
  include_x_labs = TRUE
)
}
\arguments{
\item{slick}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{pm}{Numeric value indicating the performance metric to plot from the \code{Boxplot-class} object}

\item{type}{Character string specifying the plot type.}

\item{byOM}{Logical. Facet the plots by operating model?}

\item{include_x_labs}{Logical. Include MP labels on x-axis?}
}
\value{
A \code{ggplot2} object, or a list of \code{ggplot2} objects
}
\description{
Plots boxplot, violin plot, or a combined box+violin plot for information stored
in a \linkS4class{Boxplot} object
}
\examples{
boxplot <- Boxplot(Code=c('PI1', 'PI2', 'PI3'),
                   Label=c('Performance Indicator 1',
                         'Performance Indicator 2',
                         'Performance Indicator 3'),
                   Description = c('This is the description for PI 1',
                                   'This is the description for PI 2',
                                   'This is the description for PI 3'),
                   Value=array(runif(3000), dim=c(6, 5, 4, 3)),
                   Preset=list('PI and P2'=1:2)

)

boxplot

Code(boxplot) <- c('LTY', 'STY', 'P50')
Metadata(boxplot)


# Add to `Slick` object
slick <- Slick()
Boxplot(slick) <- boxplot

plotBoxplot(slick)

plotBoxplot(slick, type='violin')

plotBoxplot(slick, type='both')
}
