% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotSpider.R
\name{plotSpider}
\alias{plotSpider}
\title{Plot \code{Spider}}
\usage{
plotSpider(
  slick,
  byOM = FALSE,
  byMP = FALSE,
  incMean = FALSE,
  incMax = FALSE,
  size.pi.label = 6,
  size.mean = 4,
  size.mp.label = 16,
  size.max.value = 4,
  size.om.title = 5,
  col.om.title = "#D6501C",
  relScale = FALSE,
  fill = byMP | byOM,
  MP_label = "Code",
  ncol = 4
)
}
\arguments{
\item{slick}{A \code{\link[=Slick-class]{Slick-class()}} object}
}
\value{
A \code{ggplot2} object
}
\description{
A Spider or Radar plot
}
\seealso{
\code{\link[=Slick-methods]{Slick-methods()}}, \code{\link[=Slick-class]{Slick-class()}}
}
