% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotTradeoff.R
\name{plotTradeoff}
\alias{plotTradeoff}
\title{Plot \code{Tradeoff}}
\usage{
plotTradeoff(
  slick,
  xPI = NULL,
  yPI = NULL,
  MP_label = "Code",
  lab_size = 6,
  point_size = 2,
  size.axis.title = 14,
  size.axis.text = 12
)
}
\arguments{
\item{slick}{A \code{\link[=Slick-class]{Slick-class()}} object}
}
\value{
A \code{ggplot2} object
}
\description{
Plot \code{Tradeoff}
}
\seealso{
\code{\link[=Slick-methods]{Slick-methods()}}, \code{\link[=Slick-class]{Slick-class()}}
}
