% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Boxplot.R,
%   R/class_Kobe.R, R/class_MPs.R, R/class_OMs.R, R/class_Quilt.R,
%   R/class_Tradeoff.R, R/class_Timeseries.R, R/class_Spider.R
\name{Preset}
\alias{Preset}
\alias{Preset<-}
\alias{Preset,Boxplot-method}
\alias{Preset<-,Boxplot-method}
\alias{Preset,Kobe-method}
\alias{Preset<-,Kobe-method}
\alias{Preset,MPs-method}
\alias{Preset<-,MPs-method}
\alias{Preset,OMs-method}
\alias{Preset<-,OMs-method}
\alias{Preset,Quilt-method}
\alias{Preset<-,Quilt-method}
\alias{Preset,Tradeoff-method}
\alias{Preset<-,Tradeoff-method}
\alias{Preset,Timeseries-method}
\alias{Preset<-,Timeseries-method}
\alias{Preset,Spider-method}
\alias{Preset<-,Spider-method}
\title{Assign or access \code{Preset} for a valid object class}
\usage{
Preset(object)

Preset(object) <- value

\S4method{Preset}{Boxplot}(object)

\S4method{Preset}{Boxplot}(object) <- value

\S4method{Preset}{Kobe}(object)

\S4method{Preset}{Kobe}(object) <- value

\S4method{Preset}{MPs}(object)

\S4method{Preset}{MPs}(object) <- value

\S4method{Preset}{OMs}(object)

\S4method{Preset}{OMs}(object) <- value

\S4method{Preset}{Quilt}(object)

\S4method{Preset}{Quilt}(object) <- value

\S4method{Preset}{Tradeoff}(object)

\S4method{Preset}{Tradeoff}(object) <- value

\S4method{Preset}{Timeseries}(object)

\S4method{Preset}{Timeseries}(object) <- value

\S4method{Preset}{Spider}(object)

\S4method{Preset}{Spider}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link[=Boxplot-class]{Boxplot-class()}}, \code{\link[=Kobe-class]{Kobe-class()}}, \code{\link[=Quilt-class]{Quilt-class()}},
\code{\link[=Spider-class]{Spider-class()}}, \code{\link[=Timeseries-class]{Timeseries-class()}}, or \code{\link[=Tradeoff-class]{Tradeoff-class()}}}

\item{value}{A \code{list}, formatted to match the class of \code{object}. See the documentation for
corresponding \code{object} class for more details.}
}
\description{
Assign or access \code{Preset} for a valid object class
}
\section{Methods (by class)}{
\itemize{
\item \code{Preset(Boxplot)}: Return \code{Preset} from a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Preset(Boxplot) <- value}: Assign \code{Preset} slot from a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Preset(Kobe)}: Return \code{Preset} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Preset(Kobe) <- value}: Assign \code{Preset} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Preset(MPs)}: Return \code{Preset} from a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Preset(MPs) <- value}: Assign \code{Preset} to a \code{\link[=MPs-class]{MPs-class()}} object

\item \code{Preset(OMs)}: Return \code{Preset} from a \code{\link[=OMs-class]{OMs-class()}} object

\item \code{Preset(OMs) <- value}: Assign \code{Preset} to an \code{OMs-class()} object

\item \code{Preset(Quilt)}: Return \code{Preset} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Preset(Quilt) <- value}: Assign \code{Preset} to a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Preset(Tradeoff)}: Return \code{Preset} from a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Preset(Tradeoff) <- value}: Assign \code{Preset} to a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Preset(Timeseries)}: Return \code{Preset} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Preset(Timeseries) <- value}: Assign \code{Preset} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Preset(Spider)}: Return \code{Preset} from a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Preset(Spider) <- value}: Assign \code{Preset} slot from a \code{\link[=Spider-class]{Spider-class()}} object

}}
