% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R
\name{Preset}
\alias{Preset}
\alias{Preset,MPs-method}
\alias{Preset<-,MPs-method}
\alias{Preset,OMs-method}
\alias{Preset<-,OMs-method}
\title{Access or modify the \code{Metadata} slot}
\usage{
Preset(x, ...)

\S4method{Preset}{MPs}(x, lang = "en")

\S4method{Preset}{MPs}(x) <- value

\S4method{Preset}{OMs}(x, lang = "en")

\S4method{Preset}{OMs}(x) <- value
}
\arguments{
\item{x}{An object with a \code{Preset} slot}

\item{...}{Additional arguments, specific to the class of object \code{x}}

\item{lang}{Character string to select language.
Either 'en', 'es', or 'fr' for English, Spanish, or French respectively.
If selected language isn't available in the object, it defaults to first language}

\item{value}{A correctly structured named list for object class \code{x}}
}
\description{
Access or modify the \code{Metadata} slot
}
\section{Methods (by class)}{
\itemize{
\item \code{Preset(MPs)}: Modify the \code{Preset} slot in an \code{\link[=MPs]{MPs()}} object

\item \code{Preset(MPs) <- value}: Assign the \code{Preset} slot in an \code{\link[=MPs]{MPs()}} object

\item \code{Preset(OMs)}: Modify the \code{Preset} slot in an \code{\link[=OMs]{OMs()}} object

\item \code{Preset(OMs) <- value}: Assign the \code{Preset} slot in an \code{\link[=OMs]{OMs()}} object

}}
