% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Slick.R
\docType{class}
\name{Slick-class}
\alias{Slick-class}
\alias{Slick}
\title{Create a \code{Slick} object}
\usage{
Slick()
}
\arguments{
\item{Title}{Title for the \code{Slick} object. A character string. For multiple languages,
use a named list with names: \code{en}, \code{es}, \code{fr} for the three supported languages.}

\item{Subtitle}{Subtitle for the \code{Slick} object. A character string or a named list with
languages: \code{en}, \code{es}, \code{fr}}

\item{Date}{Either an object of class \code{Date} or character string.
Date the Slick object was created. Defaults to \code{Sys.Date()}}

\item{Author}{A character vector with name(s) of author(s) for the \code{Slick} object}

\item{Email}{A character vector with email addresses for the author(s). Optional.
Must be same length as \code{Author}}

\item{Institution}{A character vector with institution details for the author(s). Optional.
Must be same length as \code{Author}}

\item{Introduction}{Introduction text for the \code{Slick} object. Supports all markdown formatting.
For multiple languages, use a named list with names: \code{en}, \code{es}, \code{fr} for the
three supported languages.}

\item{MPs}{An object of class \code{\link[=MPs]{MPs()}}}

\item{OMs}{An object of class \code{\link[=OMs]{OMs()}}}

\item{Boxplot}{An object of class \code{\link[=Boxplot]{Boxplot()}}}

\item{Kobe}{An object of class \code{\link[=Kobe]{Kobe()}}}

\item{Quilt}{An object of class \code{\link[=Quilt]{Quilt()}}}

\item{Spider}{An object of class \code{\link[=Spider]{Spider()}}}

\item{Timeseries}{An object of class \code{\link[=Timeseries]{Timeseries()}}}

\item{Tradeoff}{An object of class \code{\link[=Tradeoff]{Tradeoff()}}}
}
\description{
Each slot of the \code{Slick} object can be accessed or modified by the corresponding
accessor function. See Examples sections below for more details.
}
\examples{
mySlick <- Slick()
Title(mySlick) <- 'This is the title'#'
Subtitle(mySlick) <- 'This is the subtitle'
Date(mySlick) <- Sys.Date()

Title(mySlick)
Subtitle(mySlick)
Date(mySlick)


}
\author{
A. Hordyk
}
