% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Slick.R
\docType{class}
\name{Slick-class}
\alias{Slick-class}
\alias{Slick}
\title{Class \code{'Slick'}}
\usage{
Slick() # creates an new `Slick` object

Slick(Title, Subtitle, Introduction)
}
\arguments{
\item{Title}{Title for the Slick object. A character string. For multiple languages,
use a named list with names: \code{en}, \code{es}, \code{fr} for the three supported languages.}

\item{Subtitle}{Subtitle for the Slick object. A character string or a named list with
languages: \code{en}, \code{es}, \code{fr}}

\item{Introduction}{Introduction text. A list where each list element is a
separate paragraph.}
}
\description{
A \code{Slick} Data object for uploading into the Slick App.
}
\author{
A. Hordyk
}
