% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSlick.R
\name{Slick}
\alias{Slick}
\title{Run a Shiny Application}
\usage{
Slick(app = "Slick", ...)
}
\description{
\code{Slick} runs the Slick Shiny Application
}
\references{
Modified from Deal Attali's code: \url{http://deanattali.com/2015/04/21/r-package-shiny-app/}
}
