% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Timeseries.R,
%   R/class_Slick.R
\docType{methods}
\name{Timeseries}
\alias{Timeseries}
\alias{Timeseries<-}
\alias{Timeseries,missing-method}
\alias{Timeseries,character-method}
\alias{Timeseries,list-method}
\alias{Timeseries,Slick-method}
\alias{Timeseries<-,Slick-method}
\title{Methods for Creating, Accessing and Assigning \code{Timeseries} objects}
\usage{
Timeseries(
  Code = "",
  Label = "",
  Description = "",
  Time = numeric(),
  TimeNow = numeric(),
  TimeLab = "Year",
  Value = array(),
  Preset = list(),
  Target = NULL,
  Limit = NULL
)

Timeseries(Slick) <- value

\S4method{Timeseries}{missing}()

\S4method{Timeseries}{character}(
  Code = "",
  Label = "",
  Description = "",
  Time = numeric(),
  TimeNow = numeric(),
  TimeLab = "Year",
  Value = array(),
  Preset = list(),
  Target = NULL,
  Limit = NULL
)

\S4method{Timeseries}{list}(
  Code = "",
  Label = "",
  Description = "",
  Time = numeric(),
  TimeNow = numeric(),
  TimeLab = "Year",
  Value = array(),
  Preset = list(),
  Target = NULL,
  Limit = NULL
)

\S4method{Timeseries}{Slick}(Code)

\S4method{Timeseries}{Slick}(Slick) <- value
}
\arguments{
\item{Code}{A \emph{short} code for the Performance Indicators for this object.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Label}{A short label for the Performance Indicators for this object. Used to label axes on charts.
Can be longer than \code{Code} but recommended to keep short as possible so it shows clearly in plots and tables.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Description}{A description for the Performance Indicators for this object.
Can include Markdown, see \code{Examples}.
A character string length \code{nPI} or a named list for multi-language support. See \code{Details}}

\item{Time}{A numeric vector with values for the historical and projection time-steps. Must
match length \code{nTS} in \code{Value}}

\item{TimeNow}{A numeric value matching the last historical timestep in \code{Time}}

\item{TimeLab}{Character string length 1. Name of the time step (e.g., 'Year'). Will be used as the label in the \verb{Kobe Time} plot. Use a named list for
multiple languages.}

\item{Value}{A numeric array with the stochastic performance indicator values for each
simulation (sim), operating model (OM), management procedure (MP),  performance indicator (PI),
and historical + projection timestep (nTS).
Dimensions: c(\code{nsim}, \code{nOM}, \code{nMP}, \code{nPI}, \code{nTS})}

\item{Preset}{An optional named list for the preset buttons in the \code{\link[=App]{App()}}. The name of the list
element will appear as a button in the \code{\link[=App]{App()}}.}

\item{Target}{Numeric vector length \code{nPI} with the target value for the PIs.}

\item{Limit}{Numeric vector length \code{nPI} with the limit value for the PIs.}

\item{Slick}{A \code{\link[=Slick-class]{Slick-class()}} object}

\item{value}{A \code{\link[=Timeseries-class]{Timeseries-class()}} object}
}
\description{
An object of class \code{Timeseries} contains information for the Time Series chart.
The \code{Timeseries} function is used both to create and modify an \code{\link[=Timeseries-class]{Timeseries-class()}} object,
and to access and assign \code{Timeseries} for an object of class \code{\link[=Slick-class]{Slick-class()}}.
See \code{Details}.
}
\details{
Use \code{\link[=plotTimeseries]{plotTimeseries()}} to create the boxplot from the console.
\subsection{Note}{

Character strings in \code{Code}, \code{Label}, and \code{Description} must all be same length
as the number of performance indicators (\code{nPIs}) in \code{Value}

Objects of class \code{Timeseries} are created with \code{Timeseries()}
}

\subsection{Multi-Language Support}{

Text with multi-language supported can be provided as a named list. Available languages:
\itemize{
\item \code{en}: English (default)
\item \code{es}: Spanish
\item \code{fr}: French
}

Use the  \code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}} functions to access and assign the values for an
existing \code{Timeseries} object, see \code{Examples}
}
}
\section{Functions}{
\itemize{
\item \code{Timeseries(missing)}: Create an empty \code{Timeseries} object

\item \code{Timeseries(character)}: Create a populated \code{Timeseries} object

\item \code{Timeseries(list)}: Create a populated \code{Timeseries} object

\item \code{Timeseries(Slick)}: Return \code{Timeseries} from a \code{\link[=Slick-class]{Slick-class()}} object

\item \code{Timeseries(Slick) <- value}: Assign a \code{\link[=Timeseries-class]{Timeseries-class()}} object to a \code{\link[=Slick-class]{Slick-class()}} object

}}
\seealso{
\code{\link[=Timeseries-methods]{Timeseries-methods()}}, \code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}},
\code{\link[=Metadata]{Metadata()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}}, \code{\link[=plotTimeseries]{plotTimeseries()}}

\code{\link[=Code]{Code()}}, \code{\link[=Label]{Label()}}, \code{\link[=Description]{Description()}}, \code{\link[=Metadata]{Metadata()}}, \code{\link[=Value]{Value()}}, \code{\link[=Preset]{Preset()}}
}
