% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Timeseries.R, R/aa_generics.R
\name{Timeseries}
\alias{Timeseries}
\alias{Timeseries,missing,missing,missing,missing-method}
\alias{Timeseries,data.frame,missing,missing,missing-method}
\alias{Timeseries,list,missing,missing,ANY-method}
\alias{Timeseries,data.frame,data.frame,ANY,ANY-method}
\alias{Timeseries,data.frame,list,ANY,ANY-method}
\alias{Timeseries,list,data.frame,ANY,ANY-method}
\alias{Timeseries,list,list,ANY,ANY-method}
\alias{Timeseries,data.frame,data.frame,array,ANY-method}
\alias{Timeseries,data.frame,list,array,ANY-method}
\alias{Timeseries,list,data.frame,array,ANY-method}
\alias{Timeseries,list,list,array,ANY-method}
\alias{Timeseries,data.frame,list,array,list-method}
\alias{Timeseries,list,data.frame,array,list-method}
\alias{Timeseries,list,list,array,list-method}
\alias{Timeseries,data.frame,data.frame,array,list-method}
\alias{Timeseries,Slick,ANY,ANY,ANY-method}
\alias{Timeseries<-,Slick-method}
\title{\code{Timeseries} S4 class and functions}
\usage{
Timeseries(Metadata, Time, Value, Preset)

\S4method{Timeseries}{missing,missing,missing,missing}()

\S4method{Timeseries}{data.frame,missing,missing,missing}(Metadata)

\S4method{Timeseries}{list,missing,missing,ANY}(Metadata)

\S4method{Timeseries}{data.frame,data.frame,ANY,ANY}(Metadata, Time)

\S4method{Timeseries}{data.frame,list,ANY,ANY}(Metadata, Time)

\S4method{Timeseries}{list,data.frame,ANY,ANY}(Metadata, Time)

\S4method{Timeseries}{list,list,ANY,ANY}(Metadata, Time)

\S4method{Timeseries}{data.frame,data.frame,array,ANY}(Metadata, Time, Value)

\S4method{Timeseries}{data.frame,list,array,ANY}(Metadata, Time, Value)

\S4method{Timeseries}{list,data.frame,array,ANY}(Metadata, Time, Value)

\S4method{Timeseries}{list,list,array,ANY}(Metadata, Time, Value)

\S4method{Timeseries}{data.frame,list,array,list}(Metadata, Time, Value, Preset)

\S4method{Timeseries}{list,data.frame,array,list}(Metadata, Time, Value, Preset)

\S4method{Timeseries}{list,list,array,list}(Metadata, Time, Value, Preset)

\S4method{Timeseries}{data.frame,data.frame,array,list}(Metadata, Time, Value, Preset)

\S4method{Timeseries}{Slick,ANY,ANY,ANY}(Metadata)

\S4method{Timeseries}{Slick}(Slick) <- value
}
\arguments{
\item{Metadata}{An object of class \code{\link[=data.frame]{data.frame()}} or \code{\link[=list]{list()}} formatted for the
\code{Metadata} slot. See \code{Details}. Or an object of class \link{Slick}.}

\item{Time}{TODO}

\item{Value}{An \code{array} with the values for the \code{Timeseries} chart. See \code{Details}.
a character string specifying language (if available), or an object of class \code{\link[=MPs]{MPs()}}. See \code{Details}}

\item{Preset}{An optional named list. See \code{Details}.}

\item{Slick}{An object of class \link{Slick}}

\item{value}{An object of class \link{Timeseries}}
}
\value{
An object of class \code{Timeseries}
}
\description{
An object of class \code{Timeseries} contains information for the Time Series chart.
The \code{Timeseries} function is used both to create and modify an object of class \link{Timeseries}
and to access and assign the \code{Timeseries} slot in an object of class \link{Slick}. See \code{Details}.
}
\details{
\subsection{Creating objects of class \link{Timeseries}}{
\itemize{
\item \code{Timeseries()} creates a new object of class \link{Timeseries}
\item \code{Timeseries(object)} assigns the \link{Metadata} slot in an object of class \link{Timeseries},
where \code{object} is either a \code{data.frame} or a named \code{list}.  See \code{Metadata} section below.
}
}

\subsection{Accessing and assigning \code{Timeseries} slot in objects of class \link{Slick}}{

TODO
}

\subsection{Metadata}{

TODO
}

\subsection{Time}{

TODO
}

\subsection{Value}{

TODO
}

\subsection{Preset}{

TODO
}
}
\section{Slots}{

\describe{
\item{\code{Metadata}}{TODO}

\item{\code{Time}}{TODO}

\item{\code{Value}}{TODO}

\item{\code{Preset}}{TODO}
}}

