% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_generics.R, R/class_Boxplot.R,
%   R/class_Kobe.R, R/class_Quilt.R, R/class_Tradeoff.R, R/class_Timeseries.R,
%   R/class_Spider.R
\name{Value}
\alias{Value}
\alias{Value<-}
\alias{Value,Boxplot-method}
\alias{Value<-,Boxplot-method}
\alias{Value,Kobe-method}
\alias{Value<-,Kobe-method}
\alias{Value,Quilt-method}
\alias{Value<-,Quilt-method}
\alias{Value,Tradeoff-method}
\alias{Value<-,Tradeoff-method}
\alias{Value,Timeseries-method}
\alias{Value<-,Timeseries-method}
\alias{Value,Spider-method}
\alias{Value<-,Spider-method}
\title{Assign or access \code{Value} for a valid object class}
\usage{
Value(object)

Value(object) <- value

\S4method{Value}{Boxplot}(object)

\S4method{Value}{Boxplot}(object) <- value

\S4method{Value}{Kobe}(object)

\S4method{Value}{Kobe}(object) <- value

\S4method{Value}{Quilt}(object)

\S4method{Value}{Quilt}(object) <- value

\S4method{Value}{Tradeoff}(object)

\S4method{Value}{Tradeoff}(object) <- value

\S4method{Value}{Timeseries}(object)

\S4method{Value}{Timeseries}(object) <- value

\S4method{Value}{Spider}(object)

\S4method{Value}{Spider}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link[=Boxplot-class]{Boxplot-class()}}, \code{\link[=Kobe-class]{Kobe-class()}}, \code{\link[=Quilt-class]{Quilt-class()}},
\code{\link[=Spider-class]{Spider-class()}}, \code{\link[=Timeseries-class]{Timeseries-class()}}, or \code{\link[=Tradeoff-class]{Tradeoff-class()}}}

\item{value}{An \code{array}, formatted to match the class of \code{object}. See the documentation for
corresponding \code{object} class for more details.}
}
\description{
Assign or access \code{Value} for a valid object class
}
\section{Methods (by class)}{
\itemize{
\item \code{Value(Boxplot)}: Return \code{Value} from a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Value(Boxplot) <- value}: Assign \code{Value} to a \code{\link[=Boxplot-class]{Boxplot-class()}} object

\item \code{Value(Kobe)}: Return \code{Value} from a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Value(Kobe) <- value}: Assign \code{Value} to a \code{\link[=Kobe-class]{Kobe-class()}} object

\item \code{Value(Quilt)}: Return \code{Value} from a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Value(Quilt) <- value}: Assign \code{Value} to a \code{\link[=Quilt-class]{Quilt-class()}} object

\item \code{Value(Tradeoff)}: Return \code{Value} from a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Value(Tradeoff) <- value}: Assign \code{Value} to a \code{\link[=Tradeoff-class]{Tradeoff-class()}} object

\item \code{Value(Timeseries)}: Return \code{Value} from a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Value(Timeseries) <- value}: Assign \code{Value} to a \code{\link[=Timeseries-class]{Timeseries-class()}} object

\item \code{Value(Spider)}: Return \code{Value} from a \code{\link[=Spider-class]{Spider-class()}} object

\item \code{Value(Spider) <- value}: Assign \code{Value} to a \code{\link[=Spider-class]{Spider-class()}} object

}}
